Add validation and tree-walking logic to ReplaceRegex
This still does not have the regex-replacing logic. Transformation.Yaml now
builds with options, so the builder method has been renamed from build to
withOptions. It doesn't use these options yet, but taking Options is more
consistent with the other Yaml objects.

This change introduces a utility method for checking for required fields in
Configuration YAML. See ConfigValidationException.checkNotMissing.

Also, I removed the all_tests targets, because testing every target can be
accomplished with:

bazel test //...

Which, because we are not a giant mono-repo, does not risk running tests
unrelated to the project.

Change-Id: Ie68423a48b3fa53f7e6e8059214648f24d3bfbb0