Define destination repo YAML representation in config
This is a required field, since Copybara can't really do anything if it doesn't
have somewhere to copy to.

This doesn't have the implementation. I wanted to get this reviewed ASAP so that
danielromero@ and I are on the same page.

This is a single item, but it is conceivable we may want to support multiple
destinations in the future. If we do, then the transformations will be
different, so I decided against making this a list now.

Change-Id: I4b31325e95f3758dc4648b09a23cbeba6535e43e