Allow specifying workflow inside Workflow object in config YAML
This does not allow executing any Workflows besides the first one.
This will be accomplished in a follow-up change. We still allow
specifying destinations and origins in an "implicit" workflow but this
ability will eventually go away, since projects that only need one
workflow can easily create a list in YAML with only one element.

The first workflow will be the default when running Copybara without
an argument that specifies the workflow.

Change-Id: I2c59fed1a18c553a6c708c0ab768770d067de9f6