Add documentation for all the top level config elements
This change adds documentation for all the top level config elements
(Origins, Destinations, Transformations).

Because after this change all the elements contain a yamlName annotation,
I switched to extract that from the annotation instead of explicitly
repeating when registering. This allows us to not repeat ourselves and
avoid not updated configuration.

GerritDestination uses two options classes, so I had to modify the
markdown generator to accept an array of flags classes instead of
just one.

Fixed a bug where we would only get the setters from the top level
class and not the super class ones.

Change-Id: Id3816dad6fff251d3cdb7b2a2a037f4322803564