Allow $ and \ to be used in the literal portion of Replace
This requires some extra logic to escape \ and $. We could simplify
the code by requiring the user to escape these characters themself,
but since everything in templates outside of ${} is a literal, and
escaping \ for YAML as well as for the Replace template parser is
quite ugly and annoying to type, I think using $ to escape itself is
the most intuitive and elegant method.

Note that Pattern.quote doesn't quite work here since it uses \Q to
quote the string, which results in an actual 'Q' to be used in the
replacement.

Change-Id: I27bca4dbbcf34fe9103c2fa6d00389bd40ef3a6f