Sets Git/Gerrit author field as deprecated.
The field is still being used but will be replaced by
Workflow.authoring in the next CLs. Also, worth noting
that none of the configs submitted use the field, but
I'm trying to follow a standard workflow for changes
in configuration: deprecate -> migrate -> delete.

Also removes tests that won't be needed, in favor of enforcing
author validation in the Workflow.

Change-Id: Ie91d1c80ce349e102555177f38876f2e9f7bb63f