Git destinations show diff in the console before asking confirmation.
The diff shows all the staged changes before commit. There are no
unstaged changes before showing the diff because all the working
dir files are added to the staging area.

If user does not confirm, the workflow is aborted and the changes
are not committed and pushed.

This change includes a few changes to TestingConsole to be able
to match literal messages, and to have a more fluent API.

Change-Id: I3d3b1fcbf5595803d1b279fe7c94fd6f48370ff4