Fix failure when label happens twice in commit
This change fix an issue where we would crash if a label or a text
missinterpreted as a label appears twice.

Now we keep only the last appearance, that normally is the correct one.

Maybe in the future we support multivalue labels, etc. But this is
simple and good enough for now.

Change-Id: I3fc5fcfd155f516d8a259ae4451ce358d1b7017c