Move reversability check to workflow
This change moves the check of reversability to workflow instead of
having it in Replace. The error messages are similar (I would argue
that slightly better since we show a colorized git diff) and we
remove the complexity of Replace having to deal with roundtripable
transformations and passing the workflow option to transforms.

If this turns out to be a problem for our users we can add the
replace feature back.

Change-Id: I263522f556a2df2fd2751d08c938fc2ca509c259