Be able to reverse list of transforms
This change introduces the non-trivial reversal of transforms.

While we could easily provide this functionality in a bzl library,
it is so commonly used that it is better to provide it natively.

This requires a list of transforms and returns a list as a result.
I discarded to accept or return a single element because its usage
is marginal. And it is not that bad to wrap it in [].

Specifically it is not worth it to return a Reverse element because:

  - Prevents us for getting rid of Reverse

  - All the usages of the output or reverse accepts a list or is
    used to concat to an existing list (Except the rare case where
    you merge two transforms, one reversed).

Change-Id: Idc17fd98e0caa57076c0f87c8414d571c7ac8647