Implement core.replace for Skylark
This change adds support for Replace transformation using Skylark.
The function is registered under core (core.replace).

There are some hacks in place to convert from CVE to Skylark
EvalException that I expect we can make it better once we remove
YAML support.

When migrating the test I tried to keep as similar as the original
(Avoiding inlining of variables when possible) to minimize the
review time.

ReplaceYamlTest is just a copy of the old YAML ReplaceTest.

Change-Id: Ifd51ed8cbbb304019b064c19fcb7fb90c9a905fb