Use Skylark default overwrite to have better docs
This change uses a feature of Skylark that allows to define the
defaults programatically. Allowing to put anything as default
(A text). This allow us to set glob(['**']) like expressions when
appropiate instead of the non-informative and misleading None.

While this makes our code more ugly it is better for our users.

As part of this change I also make available a function to allow
creating PathMatcherBuilders without validation. This is useful
for hardcoded globs (I need this internally).

Change-Id: Ic80117503fc8f924b5faed9d89eda8a95a026f63