First change for multi-file support
This change adds support for a file wrapper that can be used to
resolve other files relative to that file (Implementation
dependant). This also will allow us to use resolvers in cases
where we don't have a filesytem.

I also made a change in to-be-deleted YamlMarkdownGenerator to
be able to compile in Intellij.

Change-Id: I5fd8b9cf0788ed66bb9033ba298db31de86e67e4