Be able to use multiple files for config
This change allows to use different files that depend on each other
for configuring Copybara. The system support two kind of labels:

  - relative to the current file 'bar/foo' gets translated to
    CURRENT/bar/foo.

  - 'absolute' paths (really are relative to the root config path.
    For example //foo/bar.

A root path heuristic is included that tries to find the root config
file by looking for the '.git' directory. IOW makes the root config
path the root of the repo. The method is protected so that we can
implement our own version internally using a similar mechanism.

For cases where we cannot find the root (Because it is not git or
other reasons) we provide a flag --config-root.

Skylark parser tries to resolve labels by adding the ".bara.sky"
suffix at the end. IOW when we have:

  load("foo/bar", "x")

We try to resolve foo/bar.bara.sky. This is the same behavior that
Skylark has in Bazel. Note that this doesn't happen for other
resources that could be loaded from the config file.

Change-Id: I55cf83f1e5da8243f2f80b0c458c9a425d8aa9ca