Configurable squash notes
This change introduces an user configurable squash notes generator.
The generator is more flexible than the old builtin one and removes
the need the need for project name for this case.

This is achieved by allowing transformations to modify the metadata
of the change. So for example the user could use it as:

  metadata.squash_notes("Super Cool project import:\n\n")

or more advances usages like:

  metadata.squash_notes(
    prefix = "Super Cool project import:\n\n",
    compact = False,
    oldest_first = True,
  )

This is the first of other message and author transformers that we will
create that are configurable by the user so that avoids having to
define their own functions for small personalization.

This change also migrates our user defined transformations for
metadata to regular transformations. We still don't expose anything
in the checkout dir. We will do in the future if we see enough
need for this.

Several refactors and code movements where needed becuase of visibility.

As part of this change include_changelist_notes is now deprecated.

Also the list of changes for squash is generated lazily.

Change-Id: I88e19efc8ddb824a593464cccd4044a73d6b4069