Allow newReader/newWriter to throw ValidationException
In the internal dest/origin implementations, some paths could be
considered invalid. This allow us to communicate the error.

I would have thrown an EvalException, but it would be non-trivial to
get a valid Location instance. Also, this is a error not related
to converting Skylark config to their representation as Java Objects
exactly (more like a contradiction between different values), so a
ValidationException seems appropriate.

Change-Id: I6cf8e19c3839ec722c2bab54820bb0fcb9b995d5