git.destination: test to exclude a directory
This test makes sure we can exclude a directory but still include the
files inside. This is important if the destination_files glob is in
the form of (e.g.) foo/bar/** - we shouldn't exclude foo/bar just
because it doesn't match the glob.

This test fails without the previous rollback.

Change-Id: Ie56b1be68e70dbd774c3981d2bc417f25a4b7816