An origin that accepts a directory as input
This change introduces an origin that can use a directory as input.

This is good for several scenarios:

  - For integration test needed internally.

  - To have basic support for version control systems/archives that
    we still don't support.

Because of the last item I added a way of defining an author and a
message for the 'change' (Even if there isn't really a change since
it is just a folder).

Change-Id: Id4bbb910753358fd5f65d29ec89f6892c4692616