metadata: A scrubber of the change message
This change allows to remove certain parts of the change message using
a regex.

The scrubber allows to use a replacement text for the matched
substrings that can use the matched groups. This allows to
scrub text outside '<public></public>' like tags. This will
probably be wrapped in a Skylark function at some point.

Change-Id: Iaa6cb08a52dc8b37f17ec9c36dca72ceb5e7fb49