Print duplicate label warning in git only if verbose
We don't have a 100% sure mechanism of detecting labels and quite
often we detect incorrect 'labels'. For example:

example = newExample()

example = new Another()

would cause the warning. In fact, this commit itself will show
the warning when importing with Copybara :)

We still show the warning if verbose flag is used.

Change-Id: If616dfa1c2fbfbbac3ea054e5b5a405e9077883c