Use git log instead of show for timestamp
Git show is a porcelain command and might show more information
than the requested format. For example with tags it shows the
tag message and author.

This makes the output unparseable since we are just expecting a
timestamp.

On the other hand git log will just output what was requested.

Change-Id: I10775b095b394c661c984fc4e11dddc4a19751a4