Avoid symlinking to ouside the checkout dir
We used to honor any symlink when copying files recursively. This is
dangerous since absolute paths or paths escaping the root of the
folder copied would be referenced instead of copied. So when later
are modified, it modifies the original ones.

This logic, while complex, allows us to detect paths that escape the
root of the folder copied and copy them as regular files.

So after this change symlinks like foo -> bar or foo/foo -> ../bar
are kept but foo -> ../baz or foo -> /absolute/path are copied.

This issue has caused several bug reports internally from customers
that got unexpected file modifications by the transforms.

Change-Id: Ib441ab8cb6ba138ab358aa944f0f39e3296d03ab