Add specific fetch function
This change introduces an specific fetch function instead of spawning
arbitrary simpleCommands from all our code base. As an added benefit
we now return the local references that changed because of the fetch:
new refs, deleted refs (if prune was used) and updated ones.

The plan is to deprecate simpleCommand from the interface as much as
possible.

As part of this change, the requirement for refspecs has been relaxed
to allow single name refs (master for example).

Change-Id: I899d9b36fcda44b8198328883b0c5e63297b7ed2