Examples annotation to validate example snippets
This change introduces a new annotation that we can add to Skylark
methods in order to associate example snippets with the function.

The mechanism allows to define a text that goes before the snippet
and an optional text that goes after.

I also added a test that executes all the snippets and validates
that they can be parsed. While this doesn't add full example
validation it covers 90% of the cases where the documentation is
outdated (Parameter renames and deprecations of fields).

We will probably need to make it more generic for Workflow, etc.
But it is good for now.

Change-Id: Ic8954b99be86426114f0564b8b6c2ca7aa5f1546