Make materialized symlinks writable
The usage of materialized symlinks is limited to folder.origin
and absolute paths symlinks. In this circumstances it is
sometimes unexpected that the file just materialized is not
writable. In particular, if we invoke Copybara inside with a
readonly tree of symlinks and we use folder.origin, we expect
the files to be writable.

If for some reason this behavior is not expected, a real origin
should be used (git.origin for example). Since folder.origin main
purpose is testing and unsupported VCSs.

Change-Id: I9ea61add4345321a1a9034785008795b93cb1a14