Remove the usage of simpleCommand()/git() for add
This change removes the usage of the more general git execution
methods in favor or an specific one. The reason is to make it
more robust, to have more validation in the future, etc.

This is part of the general effort of making simpleCommand()/git()
methods the exception and not the common use.

Change-Id: If3eac4e5839407a361206c2e99630299feb3b95b