Detect cycles in git.origin submodules
This is conservative approach where we only use the url for cycle
detection. While self-referencing to another branch is a valid
use case, this implementation is simpler and we will wait until a
valid use case from a customer happens.

This algorithm migh go down a bit more than necessary until the
urls are normalized (For example the root url usually have the
.git folder, but not in the submodules cases). This is fine since
the main purpose is to not cycle forever.

Change-Id: If27ea8191fc17d417022a3a5f2a5ab67ca31cfb7