Share ChangeQuery between GitOrigin and destination
We have feature requests that call for looking into
the destination log. To make this possible, this
change moves the log reading code to a class that
can be shared between Origin and Destination.

Makes the Visitor generic to allow type-safe handling
of the destination reference, with associated changes
for consistency.

Change-Id: I70ae07d13ec53d7129712a6ad1bb4807f572af93