Migrate --git-first-commit to a more general flag
Other destinations might benefit from knowing that this is a first
import and that it is not expected to find files/branches/commits
in the destination.

This will be used internally for another SCM.

A non-trival change was needed in a test because it was relying on
a option being mutable, so we didn't need to recreate the writer
(Baaad).

Some Cram tests needed to be updated. We still need to reactivate
them, but not in this change.

Change-Id: Ief20a1031c23973d07b3cde802b263e9ee111be2