Removes duplication in Authoring Skylark documentation.
Our internal enum had doc annotations for historical
reasons, but that documentation is duplicated from the
built in functions (overwrite, pass_thru...) and it
doesn't make sense now, because the enum cannot be used
from the config.

This reduces the confusion in the documentation.

Also renamed the enums to match the function names,
because we use the enum value in error messages and
the users need to know the built-in function that
they are using (not the enum).

Change-Id: I8ce73a260da413908cf3b58e30b88a218c0ce93d