Fix resolve for git full SHA-1
rev-parse for a full SHA-1 will always return true since there is
nothing to resolve and it is a valid SHA-1. This made our checks
for last-rev to not work (Since the destination always stores the
ref as a full SHA-1).

This error is triggered when the origin is rebased, since it
invalidates the previous history, not allowing to compute the
changes since last import. This last part is fine, but we should
show a message saying that --force can be used.

I also did some cleanup on the exceptions we throw when we don't
find a reference. I would like to make CRRE a ValidationException
but the current code makes it tricky (since we throw it in a low
level git command). I hope we can fix it soon.

Change-Id: I518ae79fb62b2f1a8dd34ec942992613234422db