More intelligent author comparison.
Git has lenient validation for the author where the email
can be empty ("Foo Bar <>" is a valid author). But also the
same author could have different names depending on the
config ("Foo Bar <foo@bar.com>" vs. "Foo B <foo@bar.com>").

With this change, we'll consider two authors the same if
they have the exact same non-empty email, but different
name. If the email is empty, then we use the name to
compare.

Change-Id: Id1f8626a4131526812884110343c9604db264380