Extracts validation into a separate class.
In this way the validation can be extended easily.

This change also includes loosing validation on the
tool command arguments. Now the validate command
accepts a workflow name (but it's not required, it
defaults to "default"), so the validation inspects
that workflow to check if it's valid in the context.

PiperOrigin-RevId: 141322082
Change-Id: I3caa8b52b6c94e637ea454f3cef9116048c4dbf5