Include commit messages of branches in migrates from GitOrigin
git.origin uses --first-parent to "flatten" history so that it appears
linear, even if the source repo uses branches heavily. In order to
avoid losing helpful commit information, this
include_branch_commit_logs setting includes the raw commit logs of all
commits on the branch.

The commit log really is raw so this has some limitations. It includes
unfiltered author information, and it repeats the message of the
"merge commit". Removing only one commit from the raw log is
non-trivial so I decided to leave it as-is.

About non-filtered authors - this is more intended for migrating a
repository in the process of changing the SoT rather than for
publishing internal/confidential branching history. If this is ever
used for the latter, this will have to be revisited.

Change-Id: I2eb3ba61a7af58395b42c3340dfe243b55c81545