Fix OOM issue on git destination commit
git.destination uses 'git diff' for checking the state of the index
in order to know if the commit would generate an empty commit. The
only thing needed there is if there are any change, not the content
of the change, so we can use --stat to only show the files changed
instead of the content of the files. This is important specially
for first migrations of big repos.

BUG=34603910

Change-Id: I251746cf148a7fded2d9539a2b0be468a48393f4