Folder origin sets writable permission on all files
This is needed once the files are copied to the workdir, otherwise
the transformations fail if the file cannot be written. Some folders
might not have all the files with write permissions, which makes
Copybara to throw AccessDeniedException.

This change does not modify the original files in the folder origin,
just the workdir files.

BUG=34969703

Change-Id: Ic5999c01f3e885a6f2b1cf396bdbb983801f2ee1