Reusing Gerrit changes
This change provides a default no op Gerrit change finder. Real
implementations should override the newGerritOptions method in the
ModuleSupplier and also override GerritOptions.newChangeFinder().
To make possible the subclassing of an options class, the lookup
in the Options class was changed. Now if an exact match was not
found, we also look for a subclass.
By default we are always trying to reuse an existing cl.

PiperOrigin-RevId: 147208342
Change-Id: I2f91faabcdac4886f33f1678e9fa4e3f9464289e