Rename Reference to Revision
Up to know we've been mixing reference with revision. For example in
git we would use it for 'master' or the SHA-1. But what we really
want is to resolve the command line to a reference using the
configuration and *then* resolve it to a particular revision. Then
the rest would work with specific revisions.

We also want to expose some information about the reference into the
workflow. For example, we should work with SHA-1 for checking in,
recording the GitOrigin-RevId, but it might be usefult to know that
this is comming from a tag called 'v1.0.1'.

In the next changes I will move Origin.resolve to use Reference and
I will polish Reference 'interface'. Also implement migrationReference.

PiperOrigin-RevId: 147629814
Change-Id: I003e9b9d4433dbeee0076cee538cc973ef5a9333