Make core.transform reversal field optional
If the field is not defined it will try to compute the reversal
from the forward transformations (Assuming they are reversible).

The usage of this is to allow to define Skylark libraries that
can return one or more transformations but that the user can
use them as a single transformation.

This avoid having weird usages like:

transformations = [
  core.replace(...),
] + some_library_function() +
[
  core.move(...),
]

Just because some_library_function wants to return two consecutive
transformations.

Instead, with this change some_library_function could wrap the
transformations with a core.transformation and return a single
transformation that can be used as:

transformations = [
    core.replace(...),
    some_library_function(),
    core.move(...),
]

Change-Id: Ie5c91679e9bd4e2d7415dba5a5fba4fafcba3a1e