Make Glob an interface and remove deprecated fields
This work is in preparation for having another Glob type that allows
the user to compose two globs as a union. We also formalize what
a Glob should provide: roots and PathMatcher when a Path is provided.

As part of this change I found a constructor that was only used by,
long time ago, deprecated fields in Workflow. So I took the
opportunity to remove it too so that we have one less glob factory
method.

I also found an accidental usage of java regex Pattern for user input.
Replaced with the re2 version.

PiperOrigin-RevId: 150279355
Change-Id: I9c976b1f354f1ea1b422bd56d35459e6514c412e