Make glob concatenable
This allows us to create the union of two globs:

   glob(['foo']) + glob (['bar'])

This would match 'foo' or 'bar'.

The specific use case that this solves is to allow including a
subdirectory again:

  glob(['foo/**'], exclude = ['foo/bar/**']) + glob(['foo/bar/baz/**'])

This allows to include all files in foo except foo/bar content but it
also includes foo/bar/baz content.

BUG=36126838

Change-Id: I5775e28047ce250ad35a15334aeb85a30ffa8719