Allow optional transformations
This mechanism allows to ignore a group of transformations. This might
make sense for multiple reasons: Temporary fix, transformations
that are exclusive, etc.

I chose to put it in core.transform since it is our effective
transform grouping mechanism and allows the user to ignore arbitrary
transformations.

Added a warning in the documentation saying that this is dangerous.

Updated documentation with latest changes.

BUG=33245255

Change-Id: Idb392b4ed904628f03bf5b259aac7ddf8fe04ce7