Replace repo.simpleCommand("log") with a specific method
This replaces all the usages of 'git log' that we have thru the code
with a common implementation that is more robust to git log format
changes.

As part of this change I discovered that 'include_branch_commit_logs'
was potentially leaking non-whitelisted authors. I fixed that. I still
don't like the current implementation. We should return a tree of
changes and let the workflow create the merge message. But this is
fine for now.

I also fixed GitRepositoryTest to use a custom GIT_AUTHOR_NAME, etc.
We need to fix the rest of the tests but I needed this for testing
git log committer.

Change-Id: Ie9728a278af9c7b4f5e8b329074aece22bbaaa83