Remove --first-parent restriction on last-rev lookup
This removes the restrition on doing --first parent when looking for
the last rev migrated. There are many cases where the label was set
in another branch and later merged. We really don't care where
the label comes from as long as we find the latest one.

I added a flag for the old behavior just in case some clients want
the old behavior.

This mechanism should be also more efficient since we rely on one
git invocation (or two in a very rare case) instead of doing one
lookup per parent.

There is a risk of a false positive, where a label is present before
the last paragraph. In that case we fallback to getting the latest
50 commits that have the label text. That should be enough, and if
it isn't the user can pass --last-rev manually.

BUG=37162996

Change-Id: If0f2a4cacc6e463ff2dd33386f327a7c2bda1c97