Fix a regression of git LogCmd with CRLF
Our old git log implementation relied on pretty output, that
automatically converts '\r\n' to '\n' in Unix systems. This is
not true when you use --format. The message is maintain as-is.

This change restores that functionality by doing a replace in
Java.

The only con is that we are not maintaining the original message,
but we already modify the message a log (for example adding the
labels.)

Change-Id: I5b3a4ec2971c1ff70b761d8891233f52b61147dc