git.gerrit_origin: Better resolve for references
This change repurposes gerrit_origin as a true gerrit origin instead
of a git origin for submitted code in Gerrit. git.origin should be
used for this case.

In particular this change allows to resolve many kind of reference
formats. From simple change number to urls or complete git references.

More importantly we store the change number as the context reference.

Change-Id: I293768ec88cdba1c48728a05ddc5e0a99b30036e