Implement check_last_rev_state field/flag
check_last_rev_state allows users to re-run Copybara for
last-rev revision and check that the generated change is
the same as the current destination state.

This validates that the change wasn't modified outside
Copybara or that new changes weren't added later to
the destination without Copybara.

The main con on this flag is that it requires an additional
migration. This flag should only be activated when users
are really unsure of the state of the destination. There was
and alternative, to use destination_files from last_rev+1..HEAD,
but that doesn't check if the last-rev commit was modified
outside of Copybara.

The test uses git repositories since it is much easier to mock
everything (And really tests the feature).

Moved GitTestUtil.java to a common testing package since now
is used from WorkflowTest too.

BUG=37472296
Change-Id: Ia9cc6374cea67b2d5bb50743d02c982c59329e46