Simplify Changes class usage
Removed Lazy Changes since it was effectively nullified by our
traversal of the history to know if we are up to date.

Use List of Change instead of Changes in forChanges and skipChanges
since we only use the list of changes to be migrated. This will
allow me to use forChanges with a list or changes before
constructing the actual Changes object so that I can filter the
changes in SQUASH mode.

Simplified forChanges semantics so that it only receives changes
in the same order. SQUASH is the only special case, so it should
be its responsability to reverse the list. Probably we could
restrict to the latest change instead of a list. But good enough
for now.

PiperOrigin-RevId: 156237881
Change-Id: Ic7326b07614a8c7a64e4ff064803f14bc0baca2a