Skip changes that don't match origin_files in SQUASH
This change makes SQUASH mode consistent with how changes are skipped
in ITERATIVE mode when they don't match origin_files.

The effect of this is that now metadata.squash_notes (or other usages
of the list of changes) only show changes that affected at least
one file in origin_files.

We use the runHelper for the computed changes so that we can use the
proper origin_files for those changes (for other implementations that
don't read the config from HEAD).

Also moved the logic to chose the current revision after the filtering
is done. Since otherwise we could be using as current a change that
is later filtered.

Personal note: DummyOrigin is too complex now. I'll refactor it to a
more sane dummy implementation soon...

BUG=38319596

Change-Id: I9b3be890fdc55a26138116a6e807a4a4c9327b1f