Refactor git related options
This change refactors git options to always use GeneralOptions for
getting the HOME directory and the directory to create repositories.

As part of this change, I also moved git.destination logic to create
local repos to the options. This will allow us in the future to have
multiple implementations of GitRepository objects.

The git.destination changes are still WIP. The ultimate goal is to
have a git.github_pr_destination, trying to reuse git.destination
logic (As much as possible).

PiperOrigin-RevId: 160523669
Change-Id: I9ae3c1d10cad8a99a2d885d27af5f8a84a1c4938