Introduce core.copy()
Copy allows to copy files as a transformation.

For relatively safe copies we create an automatic reversal that
removes the files in the destination folder. This can be erroneous
if the destination folder is not empty. But it is good compromise
between ease of usage and unexpected behavior.

Sincre core.copy and core.move are essentially the same code I chosed
to reuse the same class for now. While an abstract class might seem
like a more 'correct' approach, it makes the code more complex.

As part of this change we also expose core.remove() that allows to
remove arbitrary files. We have a basic validation that prevents
users for using it outside core.transform. As it is mean to be used
to revert core.copy. Users can still fool the validation by
wrapping it in a core.transform without a core.copy...but then
the user is conciously trying to use Copybara wrong, and we then
we don't care.

BUG=35238408

Change-Id: I819bea3992cf8b8214e4d6aa4f5638c1b3a44abe