Allow writers to maintain state in Workflows
While our workflow implementation only creates one instance of the
writer per workflow execution, other (internal) implementations
can create more writers. This allows maintaining state between
the writers.

The semantics are that the last writer requested is passed to the
next one. If dryRun is requested, the writer is not shared.

PiperOrigin-RevId: 161478564
Change-Id: I8cff8b57c33cf6f12633104661d0fb9a9918461a