Introduce metadata.map_author
Author mapping is a common pattern that currently our users need to do
with Skylark. While Skylark is good for adding unsupported/custom
transformations, it is not good when the pattern is common enough.

The mapping allows to map using just the name, the email or both. It
able to infer if it is an email, name or complete author based on:

  - Text in the form of 'a b <c@example.com>' : Complete author.

  - Contains '@' : Mail

  - Otherwise is a name.

While this fails for the general case of mail address, it is good
enough and users can always use the complete author form.

Added also a mode to fail if it cannot do the mapping (Author not
found). This is a common mistake of new authors that forget or
are not aware of the mapping and avoid not giving them attribution
or having to rewrite the history because of this.

Added an option to make the transform reversible (requires whole
authors). This allows authors to use different mail addresses
internally and externally and be able to import their external
changes.

Example usage:

    metadata.map_author{
       'Example by name <and_mail@example.com>': 'External <other@example.com>',
       'Example by name': 'External <other@example.com>',
       'example_by_mail@example.com': 'External <other@example.com>',
    }

Note that this only maps the author of the change, it doesn't change
the contents of the files (Like replacing TODOs). This is also
planned but in a different change.

Change-Id: Id2cecbfed67d3a4c1ca76a08a479be23ff11d253