Fix broken Git tests
Some tests are broken in some Linux environments because
the generated paths for the repo cache are too long.

The reason is that the Java temp directory is already long
and we also add a escaped version of the same path to the
end of it, so it's effectively the double and test throws
an error "File name too long".

Change-Id: Ifca02ba82c59b0c4136bdd90d44ecd82a1ba112a