Allow non-fast-forward pushes if fetch != push
This allows to push to branches when the user doesn't care about the
old state of the branch because it fetches from a different branch and
creates commits on top of that from a SoT.

We should use --force-with-lease instead to make it more robust, but
we don't expect concurrent writes to the branch and it is mean to
replace the whole history of the branch.

This can be used for pushing to a branch for creating/updating a
GH PR. For using like that the user would need to pass

--git-destination-push=feature_branch_name

in the CLI.

Also updated the documentation.

BUG=33387564

Change-Id: Icf5c1b4dbd7be9b0b2940e904aef6013b3af1e58