Initial implementation of git.github_pr_destination
This initial implementation is basically a wrapper of what we do in
git.destination with the behavior expected for github:

  - We rebase/rewrite all the changes newer than destination_ref
    branch (usually master).

  - Because of that we force push.

  - We fetch from destination_ref (baseline) and push to a specific
    branch for the PR (Either using groupId or using the flag)

  - This requires that origins implement extracting the groupId.
    For now, we don't have any implementation but we will have one
    soon (hash of the workflow and the following from the origin):

        * git.origin: The branch name.

        * git.gerrit_origin: Change number. We might use the topic
	  in the future for migrating multiple dependant gerrit
	  reviews using ITERATIVE mode.

        * git.github_pr_origin: The pull request number.

        * Internal SCM: Either a constant for for submitted code or
	  a the cl number for the pending change.

I also plan to add to workflow a way of creating a groupId based on
some variable using a template:

    group_id = "${username}_${workflow_name}_${context_reference}"
    hash_group_id = False/True,

For now the new destination is undocumented and the user has to
manually go to github to create a branch. In the next changes we will
use Github API to create the PR if necessary.

As part of the testing I added a way of creating customized
GitRepositories that are able to fetch/push from a different url and
do operations before/after. This example will be useful for other
implementations that need to do additional steps for fetching/pushing.

The reason for adding that extensibility in GitOptions is that we want
to be able to control this fetch/push mechanism for:

  - Normal git.origin/git.destination

  - Even for non-github/external repos we might want to do merges
    with Github PR. (Future feature).

BUG=33387564
Change-Id: Ic79048bf15027f84ae2098f7900a3eea3eea9e5b