Expose pull request number for github origin
Now we expose the number in the hidden label GITHUB_PR_NUMBER.
We only do it for urls or complete refs. We still don't do it
for single number arguments. We need to have a PR origin first.

Also added another pattern for Gerrit urls since the urls by
default end with '/'.

BUG=31362160
Change-Id: If219bd0422c6087b01f740ece7dfd70ef7455410