Updates Bazel version
This version includes a new type of Skylark event 'DEBUG'
that now we handle.

Additionally, this version includes a complete refactor
of the shell utility classes, that requires a refactor
of our usages of Command.

As part of this change I updated both the WORKSPACE
reference and the copy that we maintain in:

third_party/bazel/main/java/com/google/devtools/build/lib/shell/

This change will require updating our Copybara config
and using the same shell library internally, as it has
forked from the internal version that we're using.

BUG=64559517

Change-Id: I7b19b1df209028a781a4e2e395f47fa0b4067d4e