git.github_pr_origin boilerplate and resolve method
This first change introduces github_pr_origin, a origin for importing
GitHub Pull Requests.

Currently we can resolve these kind of references:

  - 12345: Pull Request 12345.

  - https://github.com/foo/bar/pulls/12345: We additionally validate
    that the project (foo/bar) is the one configured.

  - refs/pull/12345/head like git references.

Set COPYBARA_INTEGRATE_REVIEW label as a hidden label so that if
exposed we can use it for integrates in git.destinations.

GITHUB_PR_NUMBER label now is only set in this new origin instead of
setting it in any git origin. The label isn't currently used by
anyone and we want to restrict when this kind of metadata is set to
the specific origin. Moved the corresponding tests. 'merge' ref
coverage is lost, but we don't support it in the new origin for now.

Documentation is a bit inconsistent since we recommend to use the
still not ready origin. But I think I'll have this complete tomorrow,
so it is fine.

Created GithubUtil for commont Github urls manipulation. More things
will be added later.

BUG=37478017,65430145

Change-Id: I5dddc201a52c3821834dd812e5cb367ff09b803c