Use GIT_TERMINAL_PROMPT instead of GIT_ASKPASS
GIT_ASKPASS requires a command but GIT_TERMINAL_PROMPT allows to
simply disable it. I tested this, but maybe I used GIT_ASKPASS=true
during the manual test :(

Sadly this cannot be tested for real since it uses the tty console
directly and inside Bazel it doesn't because it is in deattached
mode.

I also took the opportunity to improve the error message for GitHub
API.

BUG=65467009
Change-Id: Id073fa6d92a79b95c2b4055314ec78d8cc73eed0