Refspec should fail harder when git not found
We were saying 'master' is not a valid refspec when the
reality is that we didn't find git. That was wrong and
confusing.

We check that git version returns a version in case of error
because if we execute:

  git check-ref-format --foo_reference

it will fail with 129 exit code and usage information. This
way we can differentiate between git not present or the ref
not being valid.

Change-Id: Ibd45bd9758d4f6045d231dce1e70a2b1f17cc818