Update TreeState between Skylark transformation calls
In order for TreeState to work, everytime we call a transformation we
need to give it an updated TreeState, so that we can check if the
next one can use the cache. This is done in Sequence for example.

SkylarkTransformation can execute arbitrary transformations. Before
this change we were reusing the same TreeState instance. So if we
executed replace twice, the second one would incorrectly assume
that the fs has been already read and that we can run the glob
inmemory.

The implementation for the fix, keeps the last TransformWork so that
a for loop like:

   for a in [very long list]:
      ctx.run(core.replace(...))

can benefit from the filesytem caching. This is also how Sequence
works.

BUG=65607455

Change-Id: I78558f0daefafacc8dbfad529add8991103bb6a5