github_pr_origin: Use complete ref names for base
While this works for regular GitRepository because we end up doing

    git fetch origin master

It doesn't work for other implementations that require unambiguous
complete reference ( 'refs/heads/master' in this case).

I tried to make a generic prefixing for refs not starting with
'refs/' but it is very tricky, because users can pass v1.0.1,
meaning a tag instead of a branch. OTOH, this fix is much simpler
and correct, since GitHub only allows to create Pull Request from
branches.

Change-Id: I546efa808d8ee66bef4083dd5c3589a87257f5b4