Refactor GerritChangeFinder
GerritChangeFinder.find method doesn't belong there. It is only used
by GerritDestination and it doesn't 'find', but generate a changeId
or reuse one. This is confusing.

This is part of a change to be able to do more kind of queries.

Removed some unit tests for find that are already covered by
GerritDestinationTest.

Also fixed some (unrelated) style issues.

PiperOrigin-RevId: 171774780
Change-Id: Ia6e96a534658cc5f684628c3c9a125f986bf7920