Revert to previous version of shell library
This is a partial revert of 2050c0f, but keeping the Bazel
update version in WORKSPACE from that commit.

It also includes moving the shell classes from package

com.google.devtools.build.lib.shell

to:

com.google.copybara.shell

To avoid conflicts with the classes that are loaded by
our Skylark dependency.

The reason for this change is to fork the shell library
that we've been using from Bazel, instead of fixing our
code every time they refactor it.

This version is identical to the one that we use
internally.

BUG=64610523

Change-Id: I33ec4ff689a1021c9e4f27900ebf13daa3e7aaac