CHANGE_REQUEST fix for non-matching origin_files
If CHANGE_REQUEST is used for a change that doesn't include any root
from origin_files, it fails with an internal error because we
expect that the lastest change is the change request one.

This is a recent regression when we added change history to
CHANGE_REQUEST.

Now we exit with a proper error, saying that the change that the
user is trying to import doesn't contain any file change matching
origin_files.

BUG=67935243

Change-Id: I8b83c9843cc72731605aea89522097cf26e4a42d