squash_notes and use_last_change honors map_author
If metadata.map_author is executed before squash_notes or
use_last_change and is configured with 'map_all_changes = True',
Copybara will honor the mapped authors in those transformations,
showing the mapped one instead of the original one.

Skylark methods renamed to be more accurate of what we are returning.
Java methods not renamed in order to minimize the change. It is not
a big deal and we can do later.

BUG=65840759

Change-Id: I1fabd856d49a723365684d476571c40f7feb3edb