Move ReadConfigFromchangeworkflow from internal
This code was originally internal, but we are trying to unify the
workflows and the mutations we allow. At some point in the future
we will also add a git implementation.

I took the oportunity to do a cleanup and allow extending the
types of commands we can have. This doesn't add much for the OSS
land but makes our internal code cleaner.

PiperOrigin-RevId: 185193265
Change-Id: I08b30d60cad42090ae07926cae220291a476ec02