Deprecate StructuredOutput and create hooks per migration
Structured output has served well to get information
about what happened on a migration, but there are
some issues that it has:

  - The builder pattern is error prone and requires
    update state triggered from other parts of the
    code (workflow resets the builder but destination
    adds information).

  - It is effectively a global variable. Making it more
    difficult what is new in each migration run (without
    diffing with previous state).

  - The destination reference is a bit vague. Splitted
    between id, type (relatively stable) and optional url.

Changed the hook mechanism to be execute per migration
instead of at the end. This is needed since sometimes
we will have a permanent side effect (git commit/push)
but the command will be killed in the middle of iterative.
Then the next execution would not see the change as pending
and wouldn't execute the hook.

While this still is not atomic, it reduces the chances of
happening.

PiperOrigin-RevId: 186524654
Change-Id: I50043d263e777a941ea7d1a843729b16b16a282d