Change-Id policy on git.gerrit_destination
Currently Copybara relies on the Gerrit server to reject a
change if it has duplicated Change-Id. This has two problems:
the error is more cryptic and doens't allow different kind
of policies when a Change-Id is already present.

This change allows to customize what to do when a Change-Id
is present/absent: Failing if not found, reusing, requiring
it always, etc.

The default is kept as it was: FAIL_IF_PRESENT. But now is
Copybara the one detecting it.

I considered introducing IGNORE mode, to don't do anything,
since our label detection algorithm is slightly different
to the Gerrit one, but I think it is fine for now.

To use the new features users can do:

    git.gerrit_destination(
        ...
        change_id_policy = 'REUSE',
        ...
    )

FIX=69907466

Change-Id: I82e6fb620335c40167ffe9e40a5c845f50afb6a4