Deal with GitHub http error codes
Our code was ignoring the error codes and returning
a general RepoException. Now we return a custom
Exception like we do for Gerrit.

Added a testcase that exercise the not found(404)
case.

Test changed so that if we don't have a response
recorded, we return a 404 like GH.

Improved the exception handling for SkylarkAction.
Now we check the exception cause and we throw the
right exception.

Changed ResponseCode map for both Gerrit and GH
api to an immutable map because it is less britte.
I had an interesting bug where the map for GH
Exception wasn't populated.

Changed CreateStatusRequest.context to be not null.
We really don't want anything to create a status
without context.

PiperOrigin-RevId: 187667124
Change-Id: Id80af6c656e3ef82ff58c98badd3e95d8baf1a8f