Use N parents for CHANGE_REQUEST_FROM_SOT
The last parent might not be exported in some circumstances (for
example if the affected code is inside a scrubbed block). This
makes the mode to fail to find the parent in the destination.

Instead, with this new approach, we will find N baselines for
the origin and we will try to find them in the destination.
The algorithm preferes the closest baseline to the change.

FIX=76109350
PiperOrigin-RevId: 190470883
Change-Id: Ifb1313ac694ec6bc7884a58a28da9878e3079ce2