Allow custom change identity
This will be used to be able to reuse destination changes
between different workflows.

The reason to expose labels is that sometimes in one workflow
the label is part of the context but in others is a label.

Note that this doesn't allow to customize the group identity.
Will be added if needed.

Change-Id: I99bdbc3acbfb4b4c3db1f3c8b678bafcae825a4e