git.gerrit_destination can push to main branches
This allows to submit directly to git main branches without using
Gerrit but generating a ChangeId. The effect of this plus
custom change identity for the workflows is that we can reuse
the destination gerrit change for pre/postsubmit.

As an alternative I considered exposing in Skylark a way of
generating your own Change-Id. But this is undesirable as users
are not interested on this and would add unnecessary complexity.

This mode makes git.gerrit_destination act as a regular
git.destination.

PiperOrigin-RevId: 190839911
Change-Id: I17faf7fdfc8215fbcf342f4b791d06c58c243b75