Fix --config-root for symlinks
A recent change (73155460b) made it --config-root folder more strict.
In particular we check that the folder is parent of the config file.

Sadly, if the folder is a symlink, it doesn't work. Since we get the
resolved config file (realpath) but we get the symlink for the root.

I noticed this issue in OSX, running our integration tests, since
/var tempfolder is a symlink to /private/var.

Added a test that works in all platforms.

PiperOrigin-RevId: 191588394
Change-Id: I56e0b50783f4b94e57f961c80f7e2d96ccadb8d6