Remove --config-root check
Remove the check that the config root has to be a parent of the
config. While normally this is sound, there are some FS
implementations that create a symlink forest and resolve to
arbitrary paths.

Furthermore, we don't really care that much. We will find the
main file and then the resolution of the dependencies will
use the roots.

Change-Id: I4bdd974aa16104fd0b05e4f6fec1178b9c9440f6