Allow commands that don't parse configuration
All our commands depend on passing at least a configuration file.
Furthermore, they do a parsing of the copy.bara.sky.

We want to be able to have commands that don't necessarily need
to load a config file. Mainly auxiliary internal commands that
are more convenient to write them inside Copybara tool (Users
don't need to install any other tool).

As part of this change I moved --help and --version to its own
commands. This allow us:

   - Simplify MainArguments

   - In the (short) future, be able to have specific help per
     command (i.e. 'copybara help migrate').

Introduced also ConfigLoaderProvider: We used to have something
similar for loading changes for specific revisions. This class
is unrelated. It is able to give a ConfigLoader only when a
command requests it, instead of creating it always. The reason
is that some internal implementations of ConfigLoader need
the config file. So we delay the creation of it until a Command
needs it.

I needed to do a hack to be able to populate the config file,
workflow name and source ref once a command parses them. The
reason is that internal code uses that information later for
reporting statistics. This has to work even in the case of
exception. And our exception mechanism is complex, so better
keep it in Main, as it is.

Removed some unit tests that were testing specific the implementation
of parsing (MainArgumentsTest). Since we already have tests in
tool_test that cover those cases much better.

MainTest.java tests have to be fixed by passing $HOME. Those tests
were passing because those checks were done before initEnvironment
call. But not anymore.

PiperOrigin-RevId: 192340824
Change-Id: Ie1d8870e7785433d0a865e7185b69e2fbfffdabf