Implement a ValidationResult for config validators
Prior to this we were using a list of console messages
which was a bad use of the console clases. First because
validation is different from  where it's going to be
printed, second because Message has more levels than the
ones required for validation (it's a different type of
message) and third because the code was too verbose.

ValidationResult encapsulates the list of validation
messages along with their level and order.

This is more convenient than passing lists around
and having to append every new message. It also provides
convenience methods to add warnings/errors with and
without formatting.

I did not use @AutoValue and @AutoValue.Builder on
purpose because it's not a good fit for this class. It
holds a list of validation messages and the code is
simpler just by implementing my own builder.

PiperOrigin-RevId: 192621230
Change-Id: I9f39718ae53c7576269a4b913c89ca810c42b254