smart_prune: Diff against a baseline run
This introduces the core of smart_prune. We run a migration for the
baseline and we compare the trees to detect the files that changed
because of the imported change.

We make the files available to the destinations so that they can
use if they want to prune the list of affected files.

In the future we might offer a more advanced pruning for affected
files sections.

The main drawback of this implementation is that it cost double
to run the migration ( 2nd checkout + transformations but not
the origin fetch or destination creation). This is not ideal but
our transformations are sort of black boxes and this approach
covers much more cases (still dynamic transforms could break
this, but it is a edge case).

BUG=33739654
Change-Id: I646a2ffa5b72b34e529e8d60fcdc7c1138d1ae0a