Make feedback context print the result to the console
This is something that probably all the users will want, and in
this way we have a default logging of everything that happens.

There is also the added benefit that if any action returns error,
feedback will throw a VE

I've also simplified the interface of validateResult() as it didn't
really need to receive BaseFunction.
BUG=78744113
PiperOrigin-RevId: 194828007
Change-Id: I75b68e0445ce404045c46402c6e5dbbf3cb020b5