Adding a list changes Gerrit api call.
When posting feedback to Gerrit, the only identifier of
the Gerrit change we have is the the patch set commit sha.
That is not enough to identify the corresponding Gerrit
change, so we need an intermediary step of getting all Gerrit
changes with that commit (in most cases should be only 1).

PiperOrigin-RevId: 194933814
Change-Id: I1305cf040845dfb5948f3fc9ba053c5534d19ebb