Expose migration effects in feedback migrations
As part of this change I had to introduce an OriginRef and
change the callers that were passing list of Change to pass
a list of OriginRefs. We were already using this from Skylark
but the change should be compatible because the field has the
same name "ref".

I've exposed two methods in Endpoint to populate origin and
destination refs because endpoints should be symmetric. We
can expose in the future raw methods for building these refs
from Skylark, but we don't need it right now.

BUG=78744113
PiperOrigin-RevId: 195158853
Change-Id: I3e64714ed2e30360b37767537173de0e88f5c384