Support pagination on GitHub API Calls
We need this for reviews, since there are instances of customers with
more than 100 reviews.

Refactored the tests to allow to pass custom headers.

I also considered putting more logic in the transports but this is
really part of the API (GitHub API defines pagination and how to
traverse the pages). Also would lead to more repeated code.

PiperOrigin-RevId: 197905508
Change-Id: I8b5629fddacafcf7fb78f5dd0d4d5b24453cfa49