Skip parsing date for dates that start with %
This happens when git was unable to parse the commit date.
While we could be more aggressive and replace any parsing
that throws DateTimeParseException, this method is more
limited in scope (it silently changes the commit date).
We want to know if we cannot parse other kind of dates.

The epoch substitution is consistent with git log output.

The test is a bit complex since now git has checks to not
allow creating commits with invalid dates.

FIXED=110171010
Change-Id: Ib3f7931e7418516874162eb8e07ac9c8f7b4a91c