Add a field for --import-noop-changes in workflow
Sometimes the user wants to make this permanent so that they don't
need to pass the flag all the times.

This also serves as a workaround for a bug that happens when
first_parent = False is used in git.origin and a noop merge happens
for a branch that has changes that modify and revert origin_files.

BUG=109929804
Change-Id: I3d3e3296e837f0e1ed4907d5bbe8577bdaf65ec1