Introduce a FileConsole and a flag to use it
Using --console-file-path, the output of the console will be also
redirected to the given file. This is useful when running Copybara
in an automated way (for example, from a Docker container). We also
extend this internally to write the output in our service.

If the file cannot be created/written, the file console will log the
error but won't fail, and will stop trying to write to the file, only
delegating from that moment on. IOW, this console is best effort.

As part of this change I'm introducing a DelegateConsole given that
we were repeating this pattern a lot.

This first implementation only flushes the writer at the very end, but
we should introduce soon a flushing mechanism (added a TODO).

BUG=111313649
BUG=111261380
PiperOrigin-RevId: 204020872
Change-Id: I6cbf97bf060c4f1dc0591e54013b30b6e08f1cca