git.destination: Consolidate writer customizations
We have two hooks for customizing git.destination writer, one
for changing the message and one for computing the effects and
logging to the console.

This change consolidate them in one in preparation for another
hook, beforePush, that will allow us to do additional checks
before pushing.

This will be used for adding an optional check in gerrit.destination
to check if the change is empty compared to the previous patchset
(if any).

Change-Id: I301a100584928588c3a5a29096bf1a004de651b3
PiperOrigin-RevId: 204817593