folder.origin symlinks between roots fix
If we have a folder.origin with a glob like:

    origin_files = glob([
        "foo/**",
	"bar/**",
    ])

And foo contains a symlink to a file in bar, we fail saying that
the the symlink escaped the roots. This is incorrect and should
be allowed.

This change fixes the issue by using the glob instead to see if
the symlinks escape. I also took the opportunity to simplify the
symlink checking code.

As part of this change, and in order to support the old behavior,
we allow foo/symlink -> '.' like symlinks (that resolve to foo).
Technically the glob foo/** doesn't include it but this would be
confusing for our users. Specially for the default origin files
["**"], that wouldn't allow this and is impossible for our users
to fix (except by excluding the symlink).

FIXED=113570475

Change-Id: Ib31d3c7da62de00d0007f15b9c747fa1302ea9c4