3way patching for git.*origin
Patching with patch.apply transformation doesn't succeed if the lines
of the file changed. 3-way diff patching is successful in more
scenarios.

This change allows to patch the checkout dir at the time of the
git checkout. Because at that point we have the git-dir, we can
apply the patch using --3way. This should be more succesful than
regular patch.apply.

As part of this change I also moved patch.apply to its own Bazel
package, since now git depends on it.

FIXED=78536811
Change-Id: I900f740588d8a0758667815597c3db993d368ae2