Revert behavior of checkout hook back to using local script
We refactored the existing hook infrastructure so it
could be used for Hg repos, too, and we added the
restriction that the hook had to be located in the
checkout dir.

This change is reverting that part to how it used to
be (script is resolved from the local filesystem) as
we don't have that requirement anymore and it simplifies
the code. The original change was done in 8773e25a.

Additionally, extracted static method into it's own
class and added utility methods in GitOriginOptions/
HgOriginOptions that make the code more readable.

PiperOrigin-RevId: 214812646
Change-Id: If159b1d7dc5811268997c80f837cef7d0222ffcb