Limit workflow mutability
Currently when using --read-config-from-change, workflows can change
arbitrarily. That means that origins, destinations and even critical
workflow options can change between each change migration in iterative
mode.

This change limits what it can change to:

   - Origin and destination glob.

   - Transformations and reverse transformations.

   - The set of config files loaded in each workflow load for a
     change. This is needed so that we can detect he changes to
     skip correctly.

In order to do that I introduce a new class ChangeMigrator in the
workflow helper that runs in the scope of one change migration.
This way we don't need to mutate he helper globs and keep the
last writer.

As part of this change I also removed the code that creates a new
writer per change. This code was tricky, since many writers maintain
state between imports (See b/69112676 for example). Instead we
require workflow to explicitly tell us what glob to use for write
and for quering destination status (so that we can use different
ones).

PiperOrigin-RevId: 215235134
Change-Id: Ie2f8d13181e9000f850ad85b61dd8c1ba3b38e13