Implement HgRevision#contexReference
As we do in other types of revisions (like Git),
if we have a configured reference provided by
the user (either in the config or as an argument)
we do return it when contextReference is called.

For example, if user runs he command as:

copybara path/to/copy.bara.sky default tip

the revision will resolve to the sha1 (globalId)
and contextReference will return 'tip'

Includes fixing the formatting of some files.

PiperOrigin-RevId: 217577898
Change-Id: I8282e62f6b4db0741e7c7ea308c382e446615f98