Add pr_branch_to_update field to github_destination
This allows us to update the head of this pr to current commit of the
main branch in GitHub only if this pull request exists. User can use
a regular string or a string with any label in changes' label list.
For instance, ${my_label}_example.

The reason behind this implementation is that presumitting changes
creates and leaves a pr open. Through this field, we can auto
merge/close this pr branch.

PiperOrigin-RevId: 220466307
Change-Id: I269f22b961081aa3c11a153c613002af1b1323e1