Relax findBaselineWithoutLabel to only care about roots
The check for including the change is too strict regarding to files.
It uses the glob but sometimes the origin change was a config
file (copy.bara.sky) stored in the origin. This can have effects
in the destination (That is why getChanges only returns files
inside the roots instead of using the glob includes/excludes).

In this particular case it is even safer, as it is only used
for CHANGE_REQUEST_FROM_SOT to find an specific label and value
in the destination.

affectedRoots moved to glob, as it is only used for globs. Still
a static field because allows us to compute the roots once.

PiperOrigin-RevId: 222314127
Change-Id: Icee41ca613aa849a3f31a2e0546fcba6355a541e