Disable Empty change description detection
This is being done too early. The destination can say that the
change was EmptyChangeException because it didn't change anything
from before commit. This is not a terminal error. ITERATIVE
continues with the next changes.

This incorrectly fails with this error as a terminal error
(ValidationException) that prevent it for continuing.

BUG=120279860
Change-Id: I874ae68f75862dd0fbe6e8839a64e6b0ec383d0e