Make fetch timeout configurable
Currently all commands have a 15m timeout limit. This is more than
enough for all of them but it is not enough for doing fetches from
remote repositories.

This change introduces a custom flag to configure a different timeout
for origin/destinations that do fetches.

Also introduced a new converter for Duration flag values. For now we
only support seconds, minutes or hours.

Fixed some visibility issues spotted by Intellij.

As alternative, I tried to make the timeout a general flag but:

  - We have a lot of commands and dependencies that use static methods.
    And all of them are fine, since we don't expect that they take 15m.
    In fact they should take much less.

  - The common/unique case is that the user wants to give Copybara more
    time for fetch. But at the same time we don't want that any command
    can use the new extended timeout.

BUG=120030093
PiperOrigin-RevId: 224189792
Change-Id: I0367c3f5bb897007e0255c2e699ee362e0bd7854