Limit query scope for listing pull requests
Currenlty git.github_pr_destination checks if a pull request already
exist for the branch about to be exported. The problem is that when
there are a lot of branches, the result is paginated and we don't
use pagination in get pull requests api (But even if we did we have
a limited number of pages we fetch).

Because of this, we fail to detect that a PR already exist and we
try to create a new one. This results in a failure from GitHub API.

This change makes the query scoped to the head that we are interested:
user and branch. Note that user here is the user of the repository
where the branch exist. For us is always the same repo (But might
change in the future).

For now we continuing querying for all PRs no matter the state. We
could query for only the ones open but I want to make sure first
that we can create a new PR with the same branch if the other is
closed or if the proper way is reopening the PR.

As part of this change I added all the possible filters to the query.

BUG=120626305

Change-Id: I58c1acdeea21f772fed581ecb62537135de4ba9d