Return None for get operations in GitHubEndpoint
Currently if the object is not found we throw GitHubApiException. The
problem is that this doesn't allow the Starlark code to act on non-found
entities. For example we might want to ignore certain failures and
continue with other refs.

This change intercepts the non found error and returns None instead.
This means that users of the API in Starlark have to do something
like this:

    commit = api.get_commit(sha)
    if not commit:
      act on not found
    happy code path

Unless they know that this cannot happen.

BUG=79919334
Change-Id: Ic3922b68283ed982aa1c937c106f0ae85d0256e8