Prevent patch auto-reversing for GNU Patch
Currenlty if a patch cannot be applied, when GNU Patch is used, it
will try to automatically reverse the patch and silently apply the
reversal.

This behavior is dangerous for Copybara usage. For example, if the
incoming change nullifies the patch (Something fixed already in
upstream), it will automatically apply the patch in reverse and
revert that fix. That is not what the user wants/expects.

This fix uses --forward to force applying the patch in only one
direction and fail otherwise. Surprisingly (or not) when we apply
the patches in reverse (-R), we also need to pass --forward so that
it doesn't try to apply it in both directions.

Test coverage added for both GNU Patch and git apply. Git apply is
not affected but is good for completeness.

FIXED=123613446

Change-Id: Ib63dc7ec5bc41bcef8cc28700b7e41a9291a599e