core.move: Delete folders when possible
Currently if we have something like:

   core.move("foo", "bar")

where foo is a directory, it doesn't delete foo. Normally this is not
a problem since git (and other SCMs) ignore directories. But this
fails when we have a file called 'foo' that we want to move it to the
root. For example:

  foo/foo

If we do:

  core.move(before = 'foo', '')

One would expect that the end result is that we have a file 'foo'. But
this instead fails because 'foo' directory wasn't deleted first. This
doesn't work even if a temporary folder is used.

For non-pure folder moves ('before' is folder and 'paths' is not used)
we don't delete empty directories recursively. Only if 'before' itself
is empty.

An additional benefit of this change is that it is much more efficient
for subfolders with lot of files, since we move the whole folder instead
of moving each individual file in Java.

FIXED=123664643
Change-Id: Ibb8a2ca40a002705ab39bd51752636ba8dc8f081