Fix a core.move bug introduced at 7e24dd84
When we try to move a folder content we were only testing if the
'after' folder existed, but not if the inner folders did.

Instead now we rely on the naive version that copies each
individual file. This still has the benefit of supporting the
case of 'foo/foo' style files issue fixed at 7e24dd8, since
we use a mid-step with a temp folder.

As part of this change I also changed the empty folder deletion
logic to be recursive. It should be like that but with the tmp
folder now is more important, as we don't want to leave temp
folders behind.

FIXED=123664643

Change-Id: If43f69a05f3c282cfc1fb7b0ebc8135fc3eaf682