Flush console in configurable interval
Currently we flush the console every x messages. This is problematic
as if a command get stuck, we see old messages instead of the latest
ones. This change flushes the console independently of having
activity.

Additionally now the DurationConverter accepts zero times (0s for
example). The only usage is fetch, and the effect would be to fail
immediately. So no much danger. If a command needs a duration with
restriction it should extend DurationConverter or check in the
caller part.

PiperOrigin-RevId: 233839381
Change-Id: I6aedc95260d04788ef3c6cbdd0e09cd48c35063c