Parse correctly Gerrit output for new/updated review
This changed recently in Gerrit. Previous format for new reviews was:

remote: New Changes:
remote:   https://some.url.google.com/1234 This is a message

and for updated ones was:

remote: Updated Changes:
remote:   https://some.url.google.com/1234 This is a message

where new format for new reviews is:

remote: SUCCESS
remote:
remote:   https://some.url.google.com/1234 This is a message [NEW]

and for updated ones is:

remote: SUCCESS
remote:
remote:   https://some.url.google.com/1234 This is a message

We try to do lenient parsing of the new format and, if it fails,
still try the old format in case we still get this from old
instances of Gerrit. Added a TODO to remove it in the future.

BUG=126679810
PiperOrigin-RevId: 236135168
Change-Id: Ib07f165a75bde468589124f8203dd32d5ccd8bb6