Add origin revisions information on empty changes
Currently we provide this in upper level messages (console) but
the effect doesn't say what change(s) where the ones failing.
This is K for good parsers of the effects, as there is an
specific origin changes field. But some internal notification
implementation doesn't support it. It is also not terrible
to have a textual representation of the error that refers
to the change.

Change-Id: I4cf3f76d926d37f1b0513ecba7389e8bc5330f23