Lazily compute git describe
Currently we are computing the 'git describe' for each change
migrated. This is not a problem for medium sized repos with
few changes migrated. But works bad for big repos with --init-history
or equivalent where lot of changes are migrated.

After this change, we compute the description lazily. When is
requested. There are two situations that could happen:

  - SQUASH: We would do at-most one call.

  - ITERATIVE: We could still do n calls. But 1-2 seconds per
    migration (that actively ask for the describe label) is not
    a big deal.

BUG=129764140
Change-Id: If71aec8fa0beab88560371314429f2d7695f62ff