git.origin: Allow importing latest tag/branch
Currently the reference migrated has to be either hardcoded in the
config or passed as a CLI parameter. This works well for branches
but not so well for tags.

This change allows to find the latest matching tag and migrate it.

The laster_version selector allows to define a regex like
core.replace, but restricting the name of the group variables names
to s0, s1, s2, etc. for strings and n0, n1, n2, etc. for numeric
ones. The difference is how we do the sorting. This allows us to
have versions like M.m.p, YYYYMMDD.XXX, etc. Out of the scope for
now is optional version elements (For example 1.0.0 vs 1.0.0rc1).

By default it ignores the CLI reference, as the job of this
mechanism is to find the latest one. There are two ways to force
this behavior:

  - --nogit-origin-version-selector: Disables any kind of version
    selector. This is mean to be used for disabling any kind of
    version selector, even the future ones.

  - Using 'force:some_ref' as the CLI reference. This will allow
    us to easily force a migration for our internal service.

In the future we'll have different types of version selectors,
for example one that migrates multiple branches/tags and not
the latest one.

BUG=129846115
Change-Id: I31f8796c6ba39222f4172ae1ee9c0ce4429e82aa