Make MetadataModule non-static
This is a long effort to remove all the instances of @SkylarkSignature
and replace them with @SkylarkCallable.

In this case we allow some fields to be positional. This is not what
we want but due to an issue in @SkylarkSignature, we were incorrectly
accepting positional values. Now we want to be compatible.

Change-Id: I08e82de5419ee524c61e222ba5d3172d58b88ceb