Fix examples test
During my refactor from static to non-static modules I realized
that our Examples test was only testing static modules. This
means that we've been not validating some examples for long time
(For example GitModule ones). In fact I discovered that we had
three (minor) errors that made it fail when test was fixed.

This change takes into account static and non-static modules
(Since we still have some and we need to figure out what to do
with global methods).

As part of the fix I added a 'minimum number of elements' test so
that we are sure we don't accidentally stop testing examples in the
future.

Change-Id: I1947570f5a8b932afeaeb3bbd4074d9f57ffc1df