Add core.filter_replace and core.replace_mapper
core.replace is limited to the kind of mappings that can do when it
has a match. The result is that users repeat several core.replace
transformations for each specific mapping value. This is very
inefficient since Copybara loads all the files and run the regex
over all the content for each transformation.

This new transformation splits the search and the replace in two:

  - core.filter_replace can search for specific strings in files.

  - Then core.replace_mapper (or just a regular map) is used to
    do the replacement over the selected text.

This has two benefits:

  - As mention above, performance and more compact code.

  - Allows to have a different type of text filtering systems
    that can use core.replace_mapper (For example specific
    language AST transformations).

BUG=131860968
Change-Id: I495a5282b9a926d0566e4a88ca7ea5848b0d5f1e