Ignore invalid symlinks inside the roots
Currently if a symlink points to an non-existant path inside the
roots, folder.origin fails while trying to set permissions. This is
fine in general as we don't want invalid symlinks and the error
shows the offending file.

There are situations where users know that there are invalid symlinks
but they don't want to add them to exclude in origin_files
(because of the maintainance cost or because they don't control the
specific workflow).

This change allows to ignore invalid symlinks in folder.origin
checkout. Flag could be used in other circumstances, but this is the
only case we found for now.

BUG=133252531
Change-Id: I2333c619a51b214db336ee5850988339464e631c