Improvements to git.github_pr_destination description
Currently git.github_pr_destination only sets the description of the
PR (title and body) when the PR is created. It doesn't update it
in next imports. Furthermore, title/body fields can only be constants.

This change adds two features:

  - Update the PR title/body even if the PR exist if
    'update_description' is set to True. For now, it is only done for
    different migration executions. In other words, for ITERATIVE mode
    we only do it once (the first change). We can revisit this if
    needed.

  - Title/body can now be a template. So if values are "Example ${foo}",
    it will look for foo label and replace in the template.

As part of this change I have to move LabelTemplate to templatetoken
package because of dependency issues. It makes sense to have this class
in this package, and in fact, I discovered that we probably should
consolidate LabelTemplate and Token/Parser classes, since they do
similar things in different ways.

FIXED=128697043

Change-Id: Iecfe0af8e5121f08d650c7571f56e9e212e21219