Protect ls-remote from logging errors
Some ls-remote operations can return potentially thousands
of refs (like for example refs/changes/* on a Gerrit
repo with a lot of changes). This can cause the underneath
logging system to crash with a single logging statement.

We've fixed internally a few instances of this by setting
an explicit limit on ls-remote operations. What we are doing
here is setting the same default for other Git commands
(4k lines) to ls-remote.

This can be changed by calling the overloaded method but it's
probably not going to be necessary. Most of the usages of
this method are for refs/changes/<change_number>/%d

And this only limits the logging, not the command output
itself, so it should be harmless.

PiperOrigin-RevId: 258356423
Change-Id: Ifad0a9529a4a623e0332e7b5fc72d97efcddae99