Sort by migration name in copybara info
The info subcommand with no migration argument, prints a
table with all the migrations in the configuration, but
does not sort the results.

For large configurations with lost of workflows this
produces an output that is hard to parse for humans.

As part of this change I've extended the testing console
to be able to skip ahead when looking for matching messages.

PiperOrigin-RevId: 259603436
Change-Id: I2ec819fc9cbc31e92774f90faa865d1b8aeda96b