Fix git integrates for external filesubjects
Currently when semi-fake integrate adds the changes that only
affect the destination (external files), it computes it using
a diff between PR head and main branch (master) head.

This is incorrect as it shows the out-of-state changes that
the original change create was synced to. This effectively
reverse non-affected changes for files not matching destination_files.

This fix instead tries to find the common parent between change head
and main branch and then computes the diff between this change and
change head. Then filters the patch to remove the files matching
destination_files (Since it is already applied and internal version
should be applied) and applies that patch instead.

The new algorithm is guarded with a temporary feature flag just in
case it has to be disabled quickly.

Change-Id: I9c82c3ae3e451bf4f0f4937ccdf2fcc7fe575693