Be able to enable/disable codepaths without code changes
Currently when we add a new codepath or rewrite some code, we need to
define a custom flag or not guard the new code at all.

This introduces a general flag for guarding this new code paths that
is very cheap to use: Just call GeneralOptions.isTemporaryFeature("foo")
This can be used either to enable something or to disable something.

This flag shouldn't be used for user facing flags. Create a flag for
this. The exception is if after the fact we have to tell some user
to use --temporary-features noFoo because something is not working. But
a rollback/fix should be the top priority.

PiperOrigin-RevId: 263210335
Change-Id: I15bbb9461fb39e8a6655ffbc1da1654d6d188b2d