Fix submodules reference resolution
Submodules fails to resolve sha1 if not in master because we fetch
only the branch from the submodule config, which is master by default.

The fix is to fetch refs/heads/* to the local repo, so the object
can be resolved. It's a bit slower than the previous version, but
safer.

BUG=140812509
PiperOrigin-RevId: 268292543
Change-Id: I34ad417fa56dee24d6debb62d36131f580ba53e4