Git Origin only checkouts specific files/directories
This will do a faster checkout for big repositories since we don't have to checkout all of files. It will only checkout those directories that matches the glob.

This aims to reduce memory usage for big repos and improve efficiency during checkout of a large repository.

Additionally, add flag experiment-checkout-affected-files to guard this feature

For internal usage

PiperOrigin-RevId: 280029618
Change-Id: If9e39e4309e1a34375b19a9f8e95c1c0160c8b6d