Add an api to read files from the destination
This is only implemented for git and works for a single file's content as well as for copying a glob.

This reads from the current iteration's baseline of the push branch, building upon a change provided by malcon@. Future work will be to package this in a traditional transform, in addition to dynamic api calls.

PiperOrigin-RevId: 292239938
Change-Id: Ic7b5de37d7c9e68c539ed01b7dd4bbfbdac83488