Retry fetching integrate ref when submodule fails
git integrates fetches the PR (or Gerrit change) as part of doing the
integrate. This is used to include the files that are only in the
external change (PR) and to do a merge (if fake/semi-fake merge).

When the repo has submodules and depending on the current git-dir and
worktree status (almost sure a new module in HEAD vs the PR) it will
fail the fetch with a "Could not access submodule" error. But it will
not affect the main repo fetch. A subsequent fetch will not fail.

This change adds a second attempt for this type of error. This is not
ideal but is the only workaround I could find. Wasn't able to create
a repro test (But can repro with all the state in copybara if I delete
the local cache).

FIXED=154160415
PiperOrigin-RevId: 310543178
Change-Id: Id7d9f39e7dc78454c36cdae4bcf8012a43adcadf