copybara: create thread for Action/Transform at call
The setDynamicEnvironment method was a legacy of when StarlarkThread
was called environment and carried environmental information, but it no
longer does. The only information this function provides now is
the print handler. It has been simplified accordingly.

The thread is now created immediately prior to its point of use.

Also: the mutability used for each Starlark.call is "closed",
causing Starlark values to become frozen. This is good practice
in most cases, but not always necessary.
PiperOrigin-RevId: 310911300
Change-Id: If937aea304a5c4c0fb46a35bdfe2c47dcd059f37