Better algorithm for gerrit.destination ref options
Currently we don't parse the pushToRefsFor field that the user
give us. We try to add more fields at the end, checking if "%"
is already there or not. This makes adding new fields too
complext.

Instead we first parse and remove all the fields and we merge
them in a Multimap. A LinkedHashMultimap is used to maintain
the existing order (not big deal, but makes the tests deterministic
and it should be memory cheap).

BUG=159696870

Change-Id: Ia912e732c3d35fce7a4bab7e40088028808dd114