gerrit destination: Only use hashtag for presubmit
Currently when git.gerrit_destination does a push, it sets the
internal hashtag for reuse changes unconditionally. This results in
an error since regular git references don't accept hashtags (and
it doesn't makes sense).

To avoid that we only add the hashtag for references like refs/for/xxxx.
While we could use git.gerrit_destination(submit = True) field, near
future work is possibly going to do Gerrit pushes using the Gerrit API.
In that case we probably want to add the hashtag.

BUG=159696870
PiperOrigin-RevId: 320666879
Change-Id: Ief4ca65e21e6a372a06eb67d3335afc501e38ccc