Skip updating the existing PR for empty changes
By default, copybara pushes a change to the existing PR. And then it triggers ci labels tests. It takes a long time to finish some ci label tests. We want to avoid pushing the same change to skip triggering the ci label test as the test result will be the same as the new change.

Add allow_empty_diff at github_pr_destination for users to decide if they want the feature.
Right now, we only check the tree of the remote commit and new commit is the same. Then we skip pushing the new commit to the existing pr.
Use main as the branch to replace the default branch name

PiperOrigin-RevId: 331440012
Change-Id: If657cdace49ac2ceb67e72ae09edd9b345dbf549