Refactor GitHubUtil for supporting GH Enterprise
Moved the relevant method to a real class that has the host as a
field. Created a default instance for github.com that is used only
in GitModule (that we will parameterize later).

There are still some instances of GITHUB_COM.* that are tricky to
change. For example git.mirror one. We will change if needed.

PiperOrigin-RevId: 337175406
Change-Id: I0fa37d3519134258370011c3fd1dd5175abb977b