feat: use macros for defining external repositories
This patch adds macros that wrap around this workspace's external
dependencies:

  - `//:repositories.bzl%copybara_repositories`
  - `//:repositories.go.bzl%copybara_go_repositories`
  - `//:repositories.maven.bzl%copybara_maven_repositories`

The primary motivation for this is to enable easier loading of copybara
from within external repositories. Prior to this patch, those external
repositories would need to re-define each and every dependency listed in
this repository's WORKSPACE file. With this patch, those external
repositories need only to load and call the macros (shown above).

An empty BUILD file is also added in this patch as it is necessary in
order to refer to the files in the workspace root.

closes #73