feat: use $HOME/.local/share/copybara as the default output root
This patch changes the default output root from `$HOME/copybara` to
`$XDG_DATA_HOME/copybara`, falling back to `$HOME/.local/share/copybara`
if the former variable is not set or null. An error is thrown if `$HOME`
is unset.

This is desirable to ensure Copybara follows the XDG Base Directory
Specification [0].

[0]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

closes #151