Introduce a new git.* mode to detect primary branch
With git switching to main, the hardcoded default can be a problem.
With this, we allow testing whether the default "master" is actually
true and automatically switch to "main" if not.

Long term we hope to remove this and always use main by default.

Big caveat: this will not work if the destination is entirely empty.

This required undoing some of the recently added branch features and
changing some exceptions.

PiperOrigin-RevId: 361904188
Change-Id: Ia8b28ec4cbfd5f4e4687476abd868f66850414d9