Move Action* to its own (java) package
Feedback is an ambiguous term here as it can refer to the general
concept of moving metadata (feedback migration or hooks) or the
specific kind of migration called "feedback".

Now it is going to be used in git.mirror to define mirror mappings
using Starlark, so moving to its own package makes sense.

Action is defined as a top level part of a migration that can
have effects in the origin/destinations and that can result in
noops, errors or success.

Change-Id: I1ecfd0de723f346028c972d3e652b6cb0d5c2f58
PiperOrigin-RevId: 370912920