Move shared methods to common Action* classes
Feedback and GitMirror actions are the same and share most of
the code. Finish hooks are also similar, the only difference
is that we don't allow to return an error/noop, since they are
kind of "auxiliary" tasks we do to the migration and we don't
want to mask a migration error with a hook error. Also a hook
error shouldn't block migrations.

Change-Id: I5e2304936b8897e221572af153ef94879b060be2