Allow having Checkers in git.destination
This change allows to use a checker in git.destination. Because
git.destination knows the files that really changed in the commit,
it can filter them, making the checker potentially faster.

Allow the testing checker to validate whole directory paths instead
of single files. This brings it in parity with internal checkers.

BUG=33676440
PiperOrigin-RevId: 423898768
Change-Id: I18de500eca054863a49a0c2a2c4b1d3da75848d3