Add support for options specified in `//:.bazelrc.user`
With this change, users can specify options for Bazel in a configuration
file at `//:.bazelrc.user`. The options specified in this file will
be merged with options specified in `//:.bazelrc`. Because this
`//:.bazelrc.user` file is loaded at the end of `//:.bazelrc`, options
specified in the former will override those set in the latter.