Change description for checkers in destination
Some checkers have the ability to change description when they find
an error. This allows to provide a default like "Internal change"
instead of halting the whole migration pipeline.

git.destination checker was able to validate description (using the
doCheck for field function) but not rewrite it.

This change adds a new interface for descriptions, since it is a
very specific behavior, that checks and transform the message if
necessary.

BUG=203457591
PiperOrigin-RevId: 427072819
Change-Id: I329333619914e25af21d81257b8a265db74a68a2