Allow listing local references in git mirror
This change allows to list the references that has been fetched in the
local git repository for git.mirror. This allows to inspect the
references and act on that information (e.g. doing a conditional
merge).

The method support filtering by refspec so that only the matching ones
are returned.

BUG=225376404
PiperOrigin-RevId: 429634660
Change-Id: I7eb6125eaa477ad9c99e46bc1037558583ee3a65