Better names fore core.action that use lambdas
Before Starlark lambdas existed, the name we got using our core.action
wrapper was the impl internal name. It is not ideal but at least is a
meaningful name.

With lambdas, we we just get the name "lambda". This is not ideal for
creating libraries, where a better name would be the caller function.

We could do this also for non-lambdas (Specially if they start with package
private character "_") but not doing it for now.

PiperOrigin-RevId: 429634660
Change-Id: I29b25159ddf106495dd56b3cea2555eac98c7b5c