Use default description for commit attempt
Currently git.destination checks for empty message before commiting.
The problem is that sometimes Copybara doesn't realize that the
change is a NOOP until commit happens. This makes migrations fail
as "empty message" for changes that would be noop (for example
only internal changes).

This change moves back the validation after commit and uses a default
message for creating the commit if the message is empty. This is fine
since we will validate the message anyway if commit is successful.

Long term we want to move this check to a DescriptionChecker, since
they run after commit and can fail or transform the description.

BUG=219951153
PiperOrigin-RevId: 429634660
Change-Id: I7752505d8cc2591c8152ffaa58eed0f5b4a78151