Add starlark structs
The starlark docs mention struct as builtin, but it is
actually bazel specific and not part of the language core.
However, as users reading the starlark docs can be
confused by this inconsistency, this adds the struct()
constructor.

This is not particularly powerful, it is just a readonly
dict with struct access

Also adds self call support to the doc generator.

BUG=209956524
PiperOrigin-RevId: 443172611
Change-Id: If133ae4226c704f00b0438691fce1f5122150309