GitRepository: Support non-ascii in ls-tree
Currently ls-tree returns paths with quotes for non-ascii files.
This is problematic because later when used, the callers don't know
if they should remove the quote or the quote is part of the name.

This change uses -z flag, that separates by \0 and doesn't add
quotes.

This was the root issue for reverting f08fb8c2 in 0df39f6c.

Change-Id: Ia9bbbe615678186a721a527980b88b6f4869d74e