repositories: pinning platforms to latest version
In Bazel versions from 5.1.0 and older, there was a change [1] which
prevents copybara from compiling on Apple Silicon by default.

The solution recommended in [2] was to pin `platforms` repository to a
newer version where the constraints value for CPP toolchain could be
correctly resolved.

Without this change, we would need to use Bazel 5.0.0 or older to
compile copybara successfully on Apple Silicon.

[1]: https://github.com/bazelbuild/bazel/pull/14844
[2]: https://github.com/bazelbuild/bazel/issues/15175