Config generation class
Given an origin directory and a destination directory, it automatically
generates the origin glob for moving files from origin to destination.

The algorithm first finds the files that are similar/same in the origin
and destination (no matter the name or dir location or that the
destination is slightly different because patching). Based on this
information, we compute the minimal glob that includes the similar
files and exclude the files in the origin not matching. The heuristic
for computing the "ideal" glob uses

   number_entries_in_include * number_entries_in_excludes

IOW, penalizes globs with lot of include entries or exclude entries.

Next changes will add minimal core.move computation and destination
glob (that should be easier/less problematic)

This code can be used as part of a config generation system to generate
an initial config.

PiperOrigin-RevId: 488980657
Change-Id: I20774e54ee5ed1c89fc924a98c8ec6ee14a016ac