Enable java 11 features
Seems like the flags somehow are not passed, so adding it directly
to the shell script.

Removed all harcoded 1.8 versions in BUILD files

Removed installation of jdk thru apt as Bazel doesn't direclty use
it for building the targets as java 11. Instead we inject the
java version as flags and Bazel downloads it.

Also use one feature in a class to prove it works.

Also remove different CI build script.

Change-Id: If3a9cee3df2b5513f449cae7ef4ef5471ba3fb9b