A recursive input resolving library
This is the first change to refactor onboarding code to acommodate
for the new requirements.

This change introduces the idea of InputProviders, generic classes
that can recursively resolve Input objects using several strategies:
reading from files, looking into existing destinations, asking users,
etc.

We are adding the following concepts:

  - Input: This is a glorified String identifier of an input. It
    stores the description, default value, type & converter (used when
    asked users for a value).

  - InputProvider: This is the central interface, given an Input it
    does its custom job to find the value. Can ask recursively to
    resolve dependant Inputs. InputProviders don't declare their
    dependencies beforehand to allow dynamic resolution based on
    the particular needs of the environment.

  - InputProviderResolver: This is the entry point of the resolution. It
    contains all the InputProviders and when called to resolve an
    Input it inject itself to the providers call to allow recursive
    resolutions and detect loops. It also wraps InputProviders to
    provide the required behavior (see bellow).

  - Some auxiliary InputProviders that provide behavior to the wrapped
    InputProviders:

      + CachedInputprovider: Prevents calling the inner provider to
        resolve if it has resolved the value already.

      + AskInputProvider: First tries to resolve using the wrapped
        InputProvider, then tries to use the default and then it might
	ask the user for a value depending on the Mode.

      + PrioritizedInputProvider: If more than one InputProvider exist
        for the same Input, it tries to resolve the Input based on the
	priority. The highest priority is the largest number.

Technically we could move the InputProvider wrapping outside of the
Resolver, but I want to see how it goes like that. It also forces the
users this particular behavior (cached, asking as last resort, etc.).

Added a couple of TODOs of an IOException that probably should be
InterruptedException. We need to check how the other callers treat
the exception, but it seems that it should be Interrupted.

Fixed TestingConsole to behave like the real console when asking
for an answer and the user feeds an incorrect answer (Real console
retries asking instead of failing). Fixed tests that rely on the
old behavior.

Change-Id: I1a2909fa843d9204b891c038b5ad4f393c69771b