First version of GeneratorCmd
This change introduces the new GeneratorCmd that uses the new
abstractions for getting inputs and generating configs.

ConfigGenerators now need to declare what inputs they consume.
This is normally not a problem for regular InputProviders but
it is for MapBasedInputProvider since it dynamically provides
Inputs based on strings, and the Inputs might not be registered
when the InputProviderResolver initializes. Now we force the
generators to provide the inputs (So they are registered)
before looking at the InputProviders.

Change-Id: Ia56ae9b2278829de192af1acc738dd57697a480b