Register a fake InputProvider if no InputProvider
Currently if no input provider is registered for an input, we return
empty. While this is not incorrect, it requires to register an input
provider for each input, even the ones that we don't have a provider
and we just want to ask the user for a value.

This change registers a provider if none if found so that the console
asks the user about a value.

The registration is done on-demand instead of the constructor to
give time to the generators to "register" interest in an Input.
Otherwise Input.registeredInputs() might not have the Input registered.

Also removed some exceptions that weren't thrown in tests.

Change-Id: I3f452e785b86d4506826d16ebcafd3d781b318b5