InputProvider that calls the config gen heuristic
This change adds a new InputProvider that by using the config
generation heuristics (checking out git repo, comparing it with the
destination), can generate the git origin glob.

As part of this change we also introduce a Converter that is able
to convert any string to a Starlark object (example Globs). Because
of the dependencies that it requires, it cannot be created statically
in Inputs. Instead we inject it as utility object in
InputProviderResolver. This could be useful for parsing other snippets
of Starlark for configs in the future.

PiperOrigin-RevId: 493281543
Change-Id: Ifa56e29c18d0fa1ac0194c7c1a734daab4e29416