Allow inferred Inputs
Currently for any input that is not present, we ask the user for a
value. There are some inputs that the user might not be able to
provide a value (e.g. intermediary inputs, etc.).

This change adds inferred only Inputs, that don't ask for a value
if an InputProvide cannot provide one.

Additionally to that, cleaned a bit the interface of
InputProviderResolver. When a non-optional value is requested
to be resolved, we return the value or throw exception. Most
of the callers were doing optional.get() unconditionally anyway.
Better to have it as an exception of CannotProvideException.

PiperOrigin-RevId: 494236071
Change-Id: I7279aef2032f3b5aa388256e253bd0fb242fd0cd