Ignore allow_empty_diff if non mergeable PR
Currently when users use allow_empty_diff = False, Copybara downloads
the existing PR and compares if the tree are the same (ignoring
rebases and message changes). This works well for skipping running tests
again.

Sometimes, if the change is non-mergeable, we want to ignore the flag
and upload anyway, as it would fix the PR conflict.

This change upload a new version in the following cases:

  - Change is not mergeable because it has conflicts.

  - Change is not mergeable because tests are failing or the status
    is unknown. We use some values from
    https://docs.github.com/en/graphql/reference/enums#mergestatestatus

We log the decision in verbose to the user so that in our service we
can see the reason for the decission and debug it in case it happens.

FIXED=262264361

Change-Id: Ifbacc77d01ffaf941eb6fc93fcabba3ecfad40d4