Fix mergeableStatus check
This field is not returned when we look for PRs by branch name.
Because of that, we are forced to do an additional GH API request
to get the field.

This is not ideal but good for now. Using GraphQL would probably
be better, as we use just a couple of fields from the PR. I swapped
the if condition since we want to avoid going to GH again if possible.

I removed a temporary flag since the feature has been there for
a while.

BUG=266841292

Change-Id: I2509fa828b3ad874ca2ba8cb3e1635a0adf0996e