Allow additional mergeable statuses to skip upload
We have some statuses that we know are safe to skip to upload when
the git diff would be empty for github_pr_destination. But there
might be other statuses where we want to skip too. This allows
users to add additional statuses to skip. The inmediate usage
is to skip when "BLOCKED". We think this would make sense in
general when this GitHub destination is not the SoT, but we want
to have the flexibility to enable it for individual use cases.

Update update docs script to use java 11 (otherwise it fails)

Change-Id: Ie21d0b1b46a2d28f930f834de0f2b1e93099ec3b