Utility class for parsing JSON from HttpResponse
The class provides special handling of MalformedJsonExceptions by copying the response content into a ByteArrayInputStream before parsing. This is needed for the new Gson libraries as the response stream is consumed before the exception is thrown.
It also attempts to remove the special Gerrit prefix line if it is found at the beginning of the stream.

Change-Id: Ic45b9cd51f14187aa93e48cd752697c154d2025f