Add list check-suits to github api
This is similar to the code for check-runs.

With this change, we will be able to upload to GitHub PR changes
when allow_empty_diff is False but some suites are failing.

Also added a string value to Conclusion so that users can use
the GitHub value (and we don't hide it)

Change-Id: Ief8921c38d2a1041b5209d217024b90eaf3876af