(4/n) Move actions to a single action field
git.mirror and core.feedback accept a list of actions as the actions
to execute. The reasoning behind this is to not have the same hack
we had in core.workflow with transformation (where we acccept both
a single transform or a list). But..

  - Transforms is different because most of the times you have more
    than one.

  - In all our internal usages we don't have a single case where
    we use more than one action. YAGNI.

  - The semantics of the list vs implementation are ambiguous:
    We say that they can be executed out of order (they are not).

  - If we need to have a list of actions at some point we can create
    a wrapper action with better semantics (like can run in parallel,
    etc.).

  - git.mirror one is particularly bad because most of the reasons
    for using multiple actions would require state between them
    (in the GitRepository being used). So out of order is not good.

  - If we need different 'actions', it is probably better to do
    in pure starlark where we are not constrained by the Action
    contract (just receives a ctx) that prevent them from sharing
    state (except git.mirror gitrepo).

Leaving for now the hooks because we have usages of it and the
cases are kind of K (real independent actions that run after
the migration).

BUG=269526710

Change-Id: I877a5e5a074d920c524cdb4e81b3ad9450b3ec7b