(6/n) First version of core.action_migration
core.action_migration is a new type of migration that tries to
generalize the features provided by core.feedback and git.mirror.

In particular it has the following features:

  - Be able to use the filesystem if declared. (feedback cannot)

  - Be able to use origin/destinations APIs, for example we could
    create a mirror that looks for open changes in GitHub (git.mirror
    cannot do this).

  - Generalizes the idea of endpoints to be able to have more than one
    endpoint (feedback can only have one, the destination).

Alternatives considered:

I considered adding the idea of capabilities, that we pass them to the
endpoint providers at the time of endpoint creation for getting a
custom endpoint that can/cannot access the filesystem. But I realized
that the migration itself would need to offer filesystem access (that
could be an enpoint itself, but for user convenience isn't) and that
most of the endpoints need a path to know where to store things. So
better to limit in one place the filesystem access.

Future work will also expose the dirFactory, to be able to create
non-workdir files (e.g. git.mirror equivalent to get access to the
cache). This sould be possible with the current architecture.

BUG=269526710

Change-Id: Ic33395ff44780e92a234007f8025386d76a1dde3