Add hashing module with CheckoutPath hash utilities.
The functions added in this change accept a CheckoutPath object for use within action migrations. String functions can be added in the future.

Block the creation of CheckoutPaths that escape the root, instead of just validating inconsistently. Removed a test because it threw an error when setting up the test instead of exercising the test. Added a test to CheckoutPath that handles the scenario instead.

BUG=270712326
PiperOrigin-RevId: 511592185
Change-Id: I7e93cba4c5e563a2a3929ce27225029f74fd4052