Adjust directory handling for autopatch
Currently, patch files are generated by diffing the
workdir to the submitted destination files. This
does not account for any changes that may have
happened to the files on import. Instead,
make a copy of the files pre-merge_import, and
diff the pre-merge_import to the post-merge_import.

If merge_import not enabled, continue as before

PiperOrigin-RevId: 518915604
Change-Id: I8682be6561f8fe132cf101c2a459d36445d65921