Add a regression for invalid UTF-8 verify_match
ede21648c7c8f2c579098171858a1017bd59643 reverted a regression introduced
in 5e9ba586ce6cba31bbabdc4e4f4586d44843712d because the function we use
for reading the string from file in core.verify_match needs to be able
to read non-utf8 files due to compatibility.

This adds a test for that case so that it doesn't happen again.

Change-Id: Ied2fe1dac9d91cf39f3965daaedb50f6143df11f