Implement the heuristic for core.move generation
The algorithm tries to find the most broad core.move that it can use
to move each file and then checks if it would move something
incorrectly. If that is the case then it moves one directory in to
do a more precise move. Iterates over this until either it finds
one directory or moves the individual file.

Added a clarifying test on how core.move works for subdirectories.

BUG=285577169

Change-Id: I06aba21053bd61ba9fff9c7f418eb1de2552b766