RAOrigin: don't use version selector for baseline resolve
Setting --force and --use-version-selector-ref to true overrides the version selector to a requested version selector (which just selects a version ref as is) and allows the supplied remotefiles.origin.resolver to be used.

This causes issues when those flags are set to false, because if the user supplies a version selector that selects the latest version, the original baseline ref will not be respected and instead the origin will just resolve a latest version found in the version list for the baseline as if it were seeking a new ref to import.

The issue is a baseline ref resolve and a new ref resolve should be treated differently.

BUG=286916256
PiperOrigin-RevId: 540680934
Change-Id: I63ccf362bb1c547d72ea99b05f8d94008ed05adb