Add RegenerateCmd to copybara
RegenerateCmd reads from two destination versions, a baseline version
with consistent patch files to diff against, and a new version with
local changes.

This change defines a new interface, PatchRegenerator, to support this
functionality.

The command requires destinations to implement the logic for re-writing
to a change. The command also supports "inference", which is the ability
to automatically determine destination versions instead of specifying
them manually on the command line, if the destination can implement it.

PiperOrigin-RevId: 543464170
Change-Id: I0f08e7034f28c3afe041e1da9b981b93bc7830c9