Add getProjectById method to GerritApi
listProjects is going to be deprecated internally, so copybara needs to
move from this method. We only use it in one place internally and we
don't use regex, so we can safely replace one with the other.

This first change just add the new method and deprecates the listing
one.

BUG=295843071

Change-Id: Ib0aafb693d035fa196cffcf8d573aca02939816a