Use import baseline if no line numbers
Currently, it is an error to attempt to regenerate without line numbers unless
the import baseline flag is enabled.

If there are no line numbers in the config, this is probably what the user is
attempting anyway, so just default to that in those cases.

Some minor typo fixes and exception handling simplification

PiperOrigin-RevId: 557817077
Change-Id: Ief71ef9f91ac68b9a48a48c1d027ebf1072933e8