Ignore symlinks in debugger
Debugger was never mean to work with symlinks. The particular
bug that trigger this issue was a symlink pointing to a non-existent
target, but in general I cannot guarantee that it works with
symlinks. The value of diffing symlink is not that big (as the
target file can be debugged), so disabling all symlinks.

Change-Id: Ibfbbe5c99a1f4f0e3fe900eacf4c736130e7eb76