cli ref used without needing `--force` flag
Previously, Copybara required the `--force` flag in order for a reference passed to cli to be fetched from a git repo. This change will allow cli references to be used by Copybara without the `--force` flag. If a cli reference is not passed, then the version_selector written in the config file is used. If there is no version_selector specified in the config, the latest version is used.

BUG=187410349
PiperOrigin-RevId: 568346195
Change-Id: I2311c3de37f943d7fb05d9ed9f435c14f0ce2dac