Fix submodule fetching for '.'
When a submodule is declared to use '.', it means that it should
track the same branch/tag name as the parent. For some reason we
used HEAD instead of matching the branch. The effect of using
HEAD is that it fetches the main remote branch (master/main), but
this would prevent it for finding commits in tags/dev-branches.

This change uses parent ref contextReference if present (e.g.
tag, branch) and if null it sets null. The effect of setting
it to null is that we fetch refs/heads/* and refs/tags/*
for the submodule, that give us high chances to find the reference.

The test was broken and was not using '.' among other things.

Change-Id: I7572cd4d3769ef6328875b023b79ed5b9263550b