Don't swallow migration error if hook Starlark fails
If we run an erroneous migration with `--dry-run`, the migration error appears in the console. However, if we run the migration without `--dry-run`, and the hook fails, we will not see the original migration error. With this change, both the original migration error and hook error will appear when the migration is ran without `--dry-run`.

BUG=254681229
PiperOrigin-RevId: 575909318
Change-Id: I79092ca85be4231a553c041a2aa1eca1b5e0fd01