Fix "setLocalConfigField" to replace fields
The "--local" flag was set after the value, which made git
consider it a value-pattern. This causes the same config
setting to be appended over and over instead of being
updated, growing to hundreds of entries after many copybara
runs.
Just fixing the order would not remove those entries, thus
switching to replace-all. Fortunately this was only is used
for fetch.prune.

PiperOrigin-RevId: 576185704
Change-Id: I9cb22db3eeb3a79a410c76396483162e7260a16a