Add a VersionSelector that uses RustVersionRequirements
This is a version selector that selects the latest version that
satisfies a cargo version requirement.
For example, "^1.2" or "1.2" would match 1.2.0 - 1.2.x
This accepts any of the supported version requirement formats,
like caret, wildcard, comparison, and tilde.

The main advantage of this vs LatestVersionSelector is that it can
handle pre-release versions correctly.

BUG=305283037
PiperOrigin-RevId: 578321367
Change-Id: I32e7040e9bea50bf4925ebbabda46834a1f42ceb