Rust: Handle downloading fuzzers with path_in_vcs defined
Certain upstream repos can host more than one crate. In these cases,
the .cargo_vcs.info.json file will tell us which directory represents
the crate's root in the upstream repo.
If this value is defined, we should ignore all other files and only
look for fuzzers in that directory.

Example: https://github.com/Byron/gitoxide has multiple crates.
If we look at the gix-trace crate, it has a .cargo_vcs_info.json file
with the value "path_in_vcs" set to "gix-trace".
This means that the crate and its fuzzers live in the gix-trace
directory of the upstream repo.
PiperOrigin-RevId: 584327901
Change-Id: I1db2c634a2397550ad6e9d2842c7952548dadcaa