An OutputStream with limit write
This change adds a new class that allows us to write to a delegate
output stream up to a limit of bytes. This allows us to limit the
amount of output that, for example, we write to the console.

The reason for this class is that CommandRunner, on verbose mode,
writes the whole output to the System.err, regardless of the
line limit set. The plan is to use an heuristic of lines * 200
or similar to calculate the number of bytes allowed and avoid
checking for new line in the outputstream.

This class allows to set a suffix to be written in the delegate
output stream after the limit has reached. This can be used to
write the byte version of '(Rest of the output skipped)' to
notice to the reader that there was more output, but we didn't
write it.

Change-Id: I8e4f505b77107c01ead28e601c17402b61a66063