Honor maxOutLogLines for verbose mode
CommandRunner logs to System.err directly in verbose mode. This
log doesn't honor maxOutLogLines, ending up flooding the user.
Honor maxOutLogLines by using the heuristic that each line is
~200 chars. Print a message in the log after that to tell the
user the rest of the output was skipped.

The caveat of this change is that for a system that expects
activity on the stream, we could end up in a situation where
the initial command ouputs > limit and then continues running
but since we don't print anything anymore, such system migth
think the command is unresponsive, killing it.

Change-Id: I1753815f82e5e196144bb89e72126272b06134a7