Optimize Git refspec checks
Currently only very basic refspecs are checked in Java. That means
that for large configs with lot of refspecs, we might be calling
git check-ref-format for too many refs. Each of these calls can
add around 50-100ms, for large configs, this is significant.

This change adds java logic to treat more cases, defaulting to
git cli if in doubt. IOW, the goal is not to be comprehensive but
to never return something as 'valid' that git CLI would say it is not.

Change-Id: I54af489d372a70f05b91bc4e9ea9780c36f27757