Fix refspec check optimization
If '*' is found in one component, don't compare against the regex
of components (As we already know it is correct and the regex doesn't
cover '*' because we special case it).

The effect of this is that we weren't really optimizing refspecs
with refs/foo/*.

Sadly the test needs to test the implementation to ensure that
we don't call git (As the real effect, ms slowness, is harder
to test).

Change-Id: Ib3d28d3404e81fa7e72ddba59c89c6403f96af8a