Skip FIFO files in copyFilesRecursively
If we try to copy over a named pipe, the operation will be blocked indefinitely. By checking if the file is a regular file or symlink first, we avoid this issue.

Also, read the symlink property from the file attributes instead of calling Files.isSymbolicLink. We don't need that extra call since we have all the attributes.

BUG=354743166
PiperOrigin-RevId: 654888880
Change-Id: I507fb23fbdc4b865f00390e71dfef708622874ae