Allow specifying custom NPM registries
Update NPM resolver and lister to support
registries other than the public NPM
instance, e.g. private registries like
Artifact Registry.

Note that the trailing slash was removed
as it is not supported by Artifact Registry.
The public NPM registry on the other does
not care whether there is a trailing slash
or not.

BUG=370485377
PiperOrigin-RevId: 680550097
Change-Id: Iaf6cfc91918c06b86cd34f51ca95e31dc0824017