Honor --force for git.mirror starlark actions
Currently when --force is used in a git.mirror() that uses starlark
for doing a push to the destination (ctx.destination_push), the flag
is not honored. Users have to pass +refs/foo/bar as the refspec
in starlark to do the push.

With this change --force is honor and applied globally to all the refs.

Change-Id: I20c67a04cff57ee28fe1e40ca44959235c027524
BUG=378963173
PiperOrigin-RevId: 696990300