Make full Git ref path available
This allows obtaining the full ref path for a resolved git revision, if available, from Info. It returns the full path as stored in the local copy of the git repo.

Note: the refspec used by Copybara is as follows:

*:refs/copybara_fetch/*
refs/tags/*:refs/tags/*

This means the returned full path will account for this refspec.

Knowing the full ref path allows us to determine if the ref is a tag something else (branch) from just the information provided in the Info object.

PiperOrigin-RevId: 700381158
Change-Id: I8ecd4d04604e343aec78a01e0765a593e5238f6f