Default autopatch directory prefix to root
The default value of autopatch_config.directory_prefix is currently "None" (as in, the string "None" - not the None value), which is really awkward - it means that, by default, copybara will attempt to write into a directory literally called "None", and fail if it doesn't exist.

This CL changes the default to something more sensible, the empty string, which will put the patches directory directly at the root of the destination.

Also, fix the documentation to use the correct default directory name ("AUTOPATCHES").

PiperOrigin-RevId: 707160892
Change-Id: I45040d52b2f4428dc69fcd6c894de1c0f5744c0d