Fix git apply when inside a parent git repo
git apply behaves differently when it is run inside some git repo, as
opposed to being used as a "classic" patch command. In some cases,
copybara intends to run git apply as the latter, but accidentally ends
up running in git repo mode because the current directory happens to be
located inside some git repo.

A typical example where this happens if when running `bazel test` from
inside a copybara git working copy - the git commands running inside the
tests accidentally pick up the copybara git repo, with hilarious
results. In practice that means the patch will be skipped because none
of the patched files are tracked in the spurious git repo.

Currently, there is code to protect against this failure mode for patch
transforms, but there is no such safeguard for the merge import code,
which produces very confusing failures instead.

This commit:

- Tricks git into thinking there is no git repo by passing
  --git-dir=/dev/null

- Removes the safeguard as the aforementioned trick makes it unnecessary

Change-Id: Ifc6e75d835f54cda4202f5408241abb4e40e42dd