Don't require a baseline if not actually needed
We shouldn't demand a baseline if the merge import process does not
actually need one. Specifically, if a consistency file is used, then the
baseline can be reconstructed from the consistency file.

This is useful when a baseline cannot be automatically obtained from
the origin. This is the case for folder origins for example, as these
don't support history. In this situation, one approach is to generate an
initial consistency file using --baseline-for-merge-import on the first
merge import run, and then rely on the consistency file for every
subsequent run. This commit makes this approach possible.

Also, this commit adds a user-visible warning in the case where we do
need a baseline, but couldn't get one. Previously we would silently
disable merge import in this case, which is confusing and arguably
dangerous.

Change-Id: Id519821679321ecd7256cb6ec3edcc70158832ec
Bug: 384688443
Bug: 381421702
Bug: 382268057