Map destination files to origin file with highest score
This fixes a bug where, in the copybara config generator, a destination file could be mapped to an origin file with a lower score than another potential match, depending on the order that the files are analyzed in.

To demonstrate, an example of the bug:

```
The origin has two files: fileA and fileB.
The origin has one file: fileC.
fileA is an exact match to fileC (highest score), fileB is a partial match to fileC but scores above the similarity threshold (30%).

When fileB is analyzed, it will replace fileA in the map since there is no check on fileA-> fileC's score, leading to an incorrect mapping.
```

This also means we can get different results depending on how the file directory is walked.

BUG=388553950
PiperOrigin-RevId: 718405949
Change-Id: Ie9930e320ae2811baa6b8eb7e40a38ca8563ee80