Implement --pinned-fixed-ref for Git origins
This allows specifying a fixed reference (e.g. SHA1) to resolve to.
The provided pinned fixed reference MUST be an ancestor of the commit referenced by the migration reference.

The main benefit is that this will preserve the change identity for the "refs/heads/main" reference. This can be useful if you plan to delay ingesting commits into another ref, but you still want to make it appear as if you are migrating "refs/heads/main" over.

As an example, you can run a migration for "refs/heads/main", but specify the flag with a value of "abcdef". Assuming "abcdef" is an ancestor of whatever "main" points to, this will result in a migration for "abcdef" but with the change identity of "refs/heads/main".

Currently, support is limited to Git origins.

PiperOrigin-RevId: 719329527
Change-Id: Icf1f72b5ebd1036e86d1499a29b0ca15f44750ef