Do not merge when origin/dest. match
Do not attempt to merge when upstream and downstream files are the same.

diff3 will not merge when the same edit happens to upstream and downstream at the same time. When the rest of the file is the same across the repos, it is safe to skip the merge.

BUG=392575734
FIXED=392575734
PiperOrigin-RevId: 723533220
Change-Id: I4bea76cdc7fb05ecd2d361c1079d3901b25e1631