Add support for GitLab API pagination
In theory, this should support both pagination methods used by GitLab: offset-based and keyset-based, since we only rely on following the next link and both methods return that. See https://docs.gitlab.com/ee/api/rest/#pagination for more info.

This is loosely modeled after our existing GitHub pagination handling, but with some simplifications:

## Flattening of the PaginatedPayload -> PaginatedList or <? extends PaginatedPayload> heirarchy
I don't see a need for that level of inheritance. GitLabApi.paginatedGet returns an immutable list of GitLabApiEntities, and that's it. The PaginatedPageList is only used as an intermediate list for constructing a complete response set. If we need to expose the list in Starlark in the future, we can just wrap it in a StarlarkList.

## Only storing the "next" link, and not the other header links provided
I don't think we'll ever use these, but fields for them were included in GitHubApi and never used.

PiperOrigin-RevId: 726574203
Change-Id: Ic173d94c241b125f2cdf339f935e4780f98ad414