Change full ref path to avoid conflicts
This can cause some conflicts if the given ref is refs/heads/main, for example. In that case, we can't write refs/copybara_fetch/refs/heads/main/copybara_full_reference.

This changes this code to save the ref to `refs/copybara_fetch/refs/heads/main_copybara_full_reference`. (note that we are not adding `copybara_full_reference` as a path component, but just as a suffix with an underscore separator.

BUG=380307584
PiperOrigin-RevId: 730997697
Change-Id: Ic3244000efc5fdedc8c4b270a26fe6b64f42700e