Add epoch version requirements
This supersedes cl/722525826

This models the "epoch" version requirements found in google3 copybara files as a separate type of Rust version requirement.

RustVersionRequirement only uses this requirement when explicitly asked to, which retains its usefulness with pure Cargo version requirements. I've made rust.bara.sky explicitly ask for this version requirement only when dealing with prereleases, but I could make it default to this. The danger there is that anyone specifying `version = 2` may start getting prereleases without asking for them.

PiperOrigin-RevId: 733435331
Change-Id: Ic96f8f07b533c383ea489b7c6828bdbfa2f52256