Compute git consistency file hashes from the FS
Previously, to generate the consistency file on git destinations, we
would run `git show` on every file and compute the hash from the output.
This is quite inefficient due to the overhead of spawning the git
command. It is much more efficient to just read the file directly from
the local checkout instead.

In my use case that involves ~55k files, this change significantly
improves copybara performance, reducing run time from ~14 minutes to ~9
minutes.

Change-Id: Ibd69f243a6b9a52ee872b2ce9da230d3059d54de