Add GitLabMrDestination and GitLabMrWriter
GitLabMrDestination: This is a destination that takes in changes from an origin, creates a new branch in the destination repo with the changes, pushes to the new branch, and creates a new GitLab MR associated with that branch.

GitLabMrWriter: This is a writer that will be used by the GitLabMrDestination to push commits/branches to GitLab, and then creating or updating a merge request via the GitLab API using the same branch.

TODO: filtering on status checks/pipeline results
BUG=393406014
PiperOrigin-RevId: 740903360
Change-Id: Ic28c7512141031b05177570f64cd487f657e919d