Add option for ignoring whitespace
Sometimes, in the config generator, the files we are trying to compare to have been formatted to meet certain style guidelines. In these cases, ignoring whitespaces leads to more accurate results for the glob matching and core.move transform heuristics.

Adding this as a flag to allow for easy enable/disable behavior.

BUG=409065048
PiperOrigin-RevId: 744862693
Change-Id: I8493ed0e86e073ea1c602c4f3ff734db846b8da7