Make GitLab endpoints use single set of credentials
Previously, these endpoints accepted both a credential issuer and a auth interceptor. Now, only a credential issuer is supplied by the user in Starlark for creating GitLabMrOrigin and GitLabMrDestination.

To ensure that the only auth used is the credential issuer, the GitLabApi object is init'ed inside of the endpoints now.

BUG=409545495
FIXED=409545495
PiperOrigin-RevId: 748004024
Change-Id: Iccb1620aaf50508ae53b3c22b2e755f25294b2ff