Allow resolving against a yanked version
This solves an issue where, if Copybara is run against a crate version that has been yanked, it can't resolve the version upstream as it filters out yanked versions.

This change allows Copybara to resolve against a yanked crate version, but not list it in the version list for potential upgrades.

BUG=423606196
FIXED=423606196
PiperOrigin-RevId: 769774385

Change-Id: I2ee7e2f0f9b8fd2da099c6f3aabeb8e264b8eff5