make copybara accept versions like ">=0.52.0, <=0.59.*"
The `ComparisonRustVersionRequirement` seems to be too strict and
fails to parse e.g. "<=0.59.*" ([from the jiff crate](
https://github.com/BurntSushi/jiff/blob/71ea43d41a3f69be6d0bec1e2ac57a012e26857f/Cargo.toml#L209C24-L209C28)). Since
it just delegates the parsing after the comparison operator, it
doesn't need to be strict about that part. Since the
`ComparisonRustVersionRequirement` type now also handles things like
`>1.0, < 2.0`, we have to make sure that
`MultipleRustVersionRequirement` is tested first.

PiperOrigin-RevId: 785998057
Change-Id: Ica0d952db84bb0b74b87d47379e32a9b1f684b37