Resolve --pinned-fixed-ref before ancestor check
When using --pinned-fixed-ref, if the SHA1 is not fetched by the origin's default refspec during the initial resolve, then the ancestor check using `git merge-base --is-ancestor` will fail with a "commit not found" error.

To prevent this, we try resolving the pinned ref before the ancestor check. If it fails, that is okay as it means it can't be found upstream, which is a legit error.

BUG=408199445
GWSQ_IGNORE: chriscampos@google.com
PiperOrigin-RevId: 820349496
Change-Id: I87001e675fd4d2683c05eb5dc86fbc09011ed886